% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SansSouci-class.R, R/SansSouciStruct-class.R
\name{nHyp}
\alias{nHyp}
\alias{nObs}
\alias{label}
\alias{pValues}
\alias{foldChanges}
\alias{thresholds}
\alias{nLeaves}
\title{\code{nHyp}: get the number of hypotheses}
\usage{
nHyp(object)

nObs(object)

label(object)

pValues(object)

foldChanges(object)

thresholds(object)

nLeaves(object)
}
\arguments{
\item{object}{An object. See individual methods for specifics}
}
\description{
\code{nHyp}: get the number of hypotheses

\code{nObs} Get the number of observations

\code{label} Get the label of hypotheses tested

\code{pValues}: get p-values

\code{foldChanges}: get fold changes

\code{thresholds}: get thresholds

\code{nLeaves}: get the number of leaves
}
