% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{as.data.tree}
\alias{as.data.tree}
\title{Convert to data tree}
\usage{
as.data.tree(C)
}
\arguments{
\item{C}{Complete tree structure}
}
\value{
A \code{data.tree::data.tree} object
}
\description{
Convert to data tree
}
\examples{

m <- 17
dd <- dyadic.from.window.size(m, s = 2, method = 2)
dt <- as.data.tree(dd$C)
plot(dt)

}
