% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaThresholdFamily.R
\name{BetaThresholdFamily}
\alias{BetaThresholdFamily}
\title{BetaThresholdFamily}
\usage{
BetaThresholdFamily(m, kMax = m)
}
\arguments{
\item{m}{The number of hypotheses tested}

\item{kMax}{For simultaneous control of (\eqn{k}-FWER for all \eqn{k \le 
k[max]}).}
}
\value{
A threshold function (on the scale of p-values) which
  achieves "balanced" joint error rate (JER) control under independence
}
\description{
Beta threshold family
}
\author{
Gilles Blanchard, Pierre Neuvial and Etienne Roquain
}
