% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biSort.R
\name{bisort}
\alias{bisort}
\title{bisort}
\usage{
bisort(mat, kMax = nrow(mat), Rcpp = TRUE)
}
\arguments{
\item{mat}{A \eqn{m} x \eqn{B} matrix of Monte-Carlo samples of test
statistics under the null hypothesis. \describe{ \item{m}{is the number of
null hypotheses tested} \item{B}{is the number of Monte-Carlo samples}}}

\item{kMax}{A scalar value between 1 and m that controls the depth of the
partial sorting of each row.}

\item{Rcpp}{If \code{TRUE} (the default), sorting operations are performed
in C++.  Otherwise, the are performed using \code{R}'s \code{base::sort}.}
}
\value{
A matrix of the same dimensions as the \code{mat} and
whose columns, then rows, have been sorted decreasingly.  For
convenience the matrix where only columns are sorted is returned
as the attribute \code{'kmaxH0'} of \code{Q}.
}
\description{
sorting a matrix decreasingly by row and then by columns
}
\details{
The output matrix is sometimes referred to as "the Q matrix", following the
notation of Meinshausen (2006).

The output matrix is sometimes referred to as "the Q
matrix", following the notation of Meinshausen (2006).
}
\examples{

m <- 1023
B <- 1e3

flavor <- c("independent", "equi-correlated", "3-factor model")[2]
rho <- 0.2
mat <- simulateGaussianNullsFromFactorModel(m, B, flavor=flavor, rho=rho)

Q <- bisort(mat, Rcpp=TRUE)
QR <- bisort(mat, Rcpp=FALSE)
identical(Q,QR)

kmaxH0 <- attr(Q, "kmaxH0")

}
\author{
Gilles Blanchard, Pierre Neuvial and Etienne Roquain
}

