% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots.R
\name{boots}
\alias{boots}
\title{Title}
\usage{
boots(
  A_est,
  f0,
  Pi_est,
  x_from,
  prob1,
  h,
  Sel_from,
  al,
  seuil,
  min_size,
  f0x_from,
  type_sel,
  num_b = 2,
  n = 2
)
}
\arguments{
\item{A_est}{matrix. The estimated transition matrix}

\item{f0}{numeric vector of size 2 (the number of states). The mean and sd of the null hypothesis (supposed to be gaussian)}

\item{Pi_est}{numeric vector of size 2 (the number of states). The estimated initial probabilities}

\item{x_from}{numeric vector. The observed variables}

\item{prob1}{numeric vector. The estimated probability for each observation to be alternatives.}

\item{h}{numeric. Size of the window for the kernel estimation of densities}

\item{Sel_from}{tibble. The different type of selection.}

\item{al}{numeric, the threshold of}

\item{seuil}{threshold of selected p-values}

\item{min_size}{the minimum size}

\item{f0x_from}{}

\item{type_sel}{}

\item{num_b}{}

\item{n}{}
}
\value{

}
\description{
Title
}
