% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boots_param}
\alias{boots_param}
\title{Title}
\usage{
boots_param(
  A_est,
  Pi_est,
  x_from,
  prob1,
  h,
  Sel_from,
  al,
  seuil,
  min_size,
  num_b,
  n,
  max_pi0,
  m0_init,
  sd0_init,
  df_init,
  norm_init,
  type_init
)
}
\arguments{
\item{A_est}{}

\item{Pi_est}{}

\item{x_from}{}

\item{prob1}{}

\item{h}{}

\item{Sel_from}{}

\item{al}{}

\item{seuil}{}

\item{min_size}{}

\item{num_b}{}

\item{n}{}

\item{max_pi0}{}

\item{m0_init}{}

\item{sd0_init}{}

\item{df_init}{}

\item{norm_init}{}
}
\value{

}
\description{
Title
}
