% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapCalibration.R
\name{bootstrap_permutation}
\alias{bootstrap_permutation}
\title{Permuted p-values using bootstrap for linear model}
\usage{
bootstrap_permutation(Y, X, C, B = 1000, replace = TRUE)
}
\arguments{
\item{Y}{Expression matrix of size $n$ observations (in row) and $D$
dimensions/mesures in columns}

\item{X}{Design matrix of size $n$ observations (in row) and $p$ variables
(in columns)}

\item{C}{Contrast matrix of size $L$ tested contrast (in row) and $p$ columns
corresponding to te tested parameters}

\item{B}{scalar, number of bootstrap}

\item{replace}{Bool. If TRUE (default) then the residuals are sampled with
replacement (i.e. a bootstrap), if FALSE then they are sampled without
replacement resulting in a permutation of the data}
}
\value{
matrix of permuted $p$-values. It is an array of dimensions
\eqn{D \times L \times B}.
}
\description{
Permuted p-values using bootstrap for linear model
}
\examples{
N = 100
p = 2
D = 2
X <- matrix(0,nrow = N, ncol = p)
X[,1] <- 1
X[,-1] <- runif(N*(p-1), min = 0, max = 3)
beta <- matrix(0, nrow = p, ncol = D)
epsilons <- matrix(rnorm(N*D), nrow = N, ncol = D)
Y <- X \%*\% beta + epsilons
C <- diag(p)
resLM <- bootstrap_permutation(Y = Y, X = X, C = C, B = 10)
}
