% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc_bounds.R
\name{bound}
\alias{bound}
\alias{bound.SansSouci}
\title{Post hoc confidence bounds on the true/false positives}
\usage{
bound(object, S, ...)

\method{bound}{SansSouci}(object, S = 1:nHyp(object), what = c("TP", "FDP"), all = FALSE, ...)
}
\arguments{
\item{object}{An object. See individual methods for specifics}

\item{S}{A subset of indices}

\item{...}{Further arguments to be passed to other methods}

\item{what}{A character vector, the names of the post hoc bounds to be
computed, among:
\itemize{
\item FP: Upper bound on the number of false positives in the 'x' most significant items
\item TP: Lower bound on the number of true positives in the 'x' most significant items
\item FDP: Upper bound on the proportion of false positives in the 'x' most significant items
\item TP: Lower bound on the proportion of true positives in the 'x' most significant items.
}

Defaults to \code{c("TP", "FDP")}}

\item{all}{A logical value: should the bounds for all ordered subsets of \code{S} be returned? If \code{FALSE} (the default), only the bound for \code{S} is returned.}
}
\value{
If \code{all} is \code{FALSE} (the default), only the value of the bound is returned. Otherwise, a \code{data.frame} is return, with |S| rows and 4 columns:
\itemize{
\item x: Number of most significant items selected
\item label: Label for the procedure, typically of the form 'refFamily(param)'
\item bound: Value of the post hoc bound
\item stat: Type of post hoc bound, as specified by argument \code{bound}.
}
}
\description{
Post hoc confidence bounds on the true/false positives
}
\examples{

# Generate Gaussian data and perform multiple tests
obj <- SansSouciSim(m = 502, rho = 0.5, n = 100, pi0 = 0.8, SNR = 3, prob = 0.5)
res <- fit(obj, B = 100, alpha = 0.1)

# post hoc bounds for all hypotheses
bound(res)

# confidence curve
cb <- bound(res, all = TRUE)
head(cb)
plot(res)

# confidence curve for a subset
S <- which(foldChanges(res) > 0.3)
plot(res, S = S)

# plot two confidence curves
res_beta <- fit(obj, B = 100, alpha = 0.1, refFamily = "Beta", K = 20)
cb_beta <- bound(res_beta, all = TRUE)

bounds <- list("Simes"= cb, 
                   "Beta" = cb_beta)
plotConfCurve(bounds, xmax = 200)

}
