% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate_jer}
\alias{calibrate_jer}
\title{Perform JER calibration}
\usage{
calibrate_jer(
  X,
  label,
  alpha,
  B = 100,
  row_test_fun = rowWelchTests,
  template = "Simes",
  k_max = nrow(X)
)
}
\arguments{
\item{X}{A matrix of \code{m} variables (hypotheses) by \code{n} observations}

\item{label}{A numeric vector of \eqn{n} values in \eqn{0, 1}, the groups of
observations on which to perform two-sample tests}

\item{alpha}{A numeric value in \verb{[0,1]}, the target (JER) risk}

\item{B}{An integer value, the number of permutations to be performed}

\item{row_test_fun}{A (vectorized) test function. Defaults to \link{rowWelchTests}.}

\item{template}{A character value, the name of a threshold family. Should be
one of "Linear", "Beta" and "Simes", or "Oracle". "Linear" and "Simes" families are
identical.}

\item{k_max}{A numeric value in \verb{[0,m]}, the length of the template}
}
\value{
A list with elements
\itemize{
\item thr: A numeric vector of length K, such that the estimated probability that
there exists an index k between 1 and K such that the k-th maximum of the
test statistics of is greater than \code{thr[k]}, is less than alpha
\item piv_stat: A vector of \code{B} pivotal statitics
\item lambda: A numeric value, the result of the calibration
}
}
\description{
Perform JER calibration
}
\details{
'calibrate0' performs single step calibration,
whereas 'calibrate' performs step-down calibration. Hence
the output of 'calibrate(..., max_steps_down = 0)' and
'calibrate0(...)' should be identical
}
\references{
Blanchard, G., Neuvial, P., & Roquain, E. (2020). Post hoc
confidence bounds on false positives using reference families. \emph{Annals of
Statistics}, 48(3), 1281-1303.
}
