% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate_simes_one_sample}
\alias{calibrate_simes_one_sample}
\title{Perform JER calibration with the Simes template for one sample.}
\usage{
calibrate_simes_one_sample(X, alpha, B = 100)
}
\arguments{
\item{X}{A matrix of \code{m} variables (hypotheses) by \code{n} observations}

\item{alpha}{A numeric value in \verb{[0,1]}, the target (JER) risk}

\item{B}{An integer value, the number of permutations to be performed}
}
\value{
A list with elements
\itemize{
\item thr: A numeric vector of length K, such that the estimated probability that
there exists an index k between 1 and K such that the k-th maximum of the
test statistics of is greater than \code{thr[k]}, is less than alpha
\item piv_stat: A vector of \code{B} pivotal statitics
\item lambda: A numeric value, the result of the calibration
}
}
\description{
Perform JER calibration with the Simes template for one sample.
}
\details{
calibrate_simes_one_sample run calibrate_jer for one sample.
Label is NULL, the template is fixed to "Simes" and k_max = nrow(X)
}
