% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JER_calibration.R
\name{calibrateJER}
\alias{calibrateJER}
\title{Calibration of joint Family-Wise Error Rate thresholds}
\usage{
calibrateJER(X, B, alpha, refFamily = c("Simes", "kFWER"), K = nrow(X),
  verbose = TRUE)
}
\arguments{
\item{X}{A matrix of \eqn{m} variables (hypotheses) by \eqn{n} observations.
The column names of X should be "0" for the first sample and "1" for the
second sample.}

\item{B}{A numeric value, the number of permutations to be performed}

\item{alpha}{Target JER level}

\item{refFamily}{A character value which can be \describe{

  \item{Simes}{The classical family of thresholds introduced by Simes (1986):
  \eqn{\alpha*k/m}. This family yields joint FWER control if the test
  statistics are positively dependent (PRDS) under H0.}

  \item{kFWER}{A family \eqn{(t_k)} calibrated so that for each k,
  \eqn{(t_k)} controls the (marginal) k-FWER.}}}

\item{K}{For JER control over \code{1:K}, ie joint control of all
\eqn{k}-FWER, \eqn{k \le K}.}

\item{verbose}{A boolean value: should extra info be printed?}
}
\value{
A list with elements: \describe{

  \item{stat}{A numeric vector of \code{m} test statistics}

  \item{thr}{A numeric vector of length \code{K}, such that the estimated
  probability that there exists an index \eqn{k} between 1 and \eqn{K} such
  that the \eqn{k}-th maximum of the test statistics of is greater than
  \eqn{thr[k]}, is less than \eqn{\alpha}}

  \item{lambda}{A numeric value, the result of the calibration} }
}
\description{
Calibration of of JER thresholds from one or two-sample tests
}
\examples{

sim <- gaussianSamples(m = 123, rho = 0.2, n = 100,
                       pi0 = 0.8, SNR = 1, prob = 0.5)
X <- sim$X
cal <- calibrateJER(X, B = 1e3, alpha = 0.2, refFamily="Simes")
cal$lambda # > alpha (whp) is rho > 0
}
\author{
Gilles Blanchard, Pierre Neuvial and Etienne Roquain
}
