% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceEnvelope.R
\name{confidenceEnvelope}
\alias{confidenceEnvelope}
\title{Confidence envelope on the true/false positives among most significant items}
\usage{
confidenceEnvelope(
  stat,
  refFamily,
  param,
  K = length(stat),
  what = c("TP", "FDP")
)
}
\arguments{
\item{stat}{A vector containing all \eqn{m} test statistics, sorted
non-increasingly}

\item{refFamily}{A character value, the reference family to be used. Should
be either "Simes" (or equivalenlty, "Linear"), "Beta", or "Oracle".}

\item{param}{A numeric value or vector of parameters for the reference family.}

\item{K}{For JER control over \code{1:K}, ie joint control of all
\eqn{k}-FWER, \eqn{k \le K}. Defaults to \eqn{m}.}

\item{what}{A character vector, the names of the post hoc bounds to be
  computed, among:

\describe{
\item{FP}{Upper bound on the number of false positives in the 'x' most significant items}
\item{TP}{Lower bound on the number of true positives in the 'x' most significant items}
\item{FDP}{Upper bound on the proportion of false positives in the 'x' most significant items}
\item{TP}{Lower bound on the proportion of true positives in the 'x' most significant items}}
Defaults to \code{c("TP", "FDP")}.}
}
\value{
A \code{matrix} with \eqn{m} rows and 5 columns: \describe{
\item{x}{Number of most significant items selected}
\item{family}{Matches input argument \code{refFamily}}
\item{param}{Matches argument \code{param}}
\item{procedure}{Label for the procedure, typically of the form '<refFamily>(<param>)'}
\item{bound}{Value of the post hoc bound}
\item{stat}{Type of post hoc bound, as specified by argument \code{bound}}
}
}
\description{
Confidence envelope on the true/false positives among most significant items
}
\details{
\code{param} should be a numeric value unless \code{refFamily ==
  "Oracle"}. In the latter case, \code{param} should be a boolean vector of
  length \eqn{m} indicating whether each null hypothesis is true or false.
}
\examples{

m <- 511
alpha <- 0.1
sim <- gaussianSamples(m = m, rho = 0.5, n = 100, pi0 = 0.8, SNR = 3, prob = 0.5)
dat <- sim$X
rwt <- rowWelchTests(dat, categ=colnames(dat), alternative = "greater")

ce <- confidenceEnvelope(rwt$statistic, refFamily = "Simes", param = alpha, what = c("TP"))


library("ggplot2")
ggplot(subset(ce, x <= 200), aes(x = x, y = bound)) +
  geom_line() + 
  facet_wrap(~ stat, scales = "free_y") + 
  labs(x = "# top genes called significant", y = "Post hoc confidence bounds")
}
\author{
Gilles Blanchard, Pierre Neuvial and Etienne Roquain
}
