% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc-bounds.R
\name{curveMaxFP_BNR2014}
\alias{curveMaxFP_BNR2014}
\alias{curveMaxFP_Mein2006}
\title{Upper bound for the number of false discoveries among most significant items}
\usage{
curveMaxFP_BNR2014(p.values, thr)

curveMaxFP_Mein2006(p.values, thr)
}
\arguments{
\item{p.values}{A vector containing all \eqn{m} p-values, sorted non-decreasingly}

\item{thr}{A vector of \eqn{K} JER-controlling thresholds, sorted non-decreasingly}
}
\value{
A vector of size \eqn{m} giving an joint upper confidence bound on
the number of false discoveries among the \eqn{k} most significant items
for all \eqn{k \in \{1\ldots m\}}.
}
\description{
Upper bound for the number of false discoveries among most significant items
}
\details{
These older implementations of 'curveMaxFP' are here for the purpose of testing that the current one yields consistent results.
}
\author{
Gilles Blanchard, Nicolas Enjalbert-Courrech, Pierre Neuvial and Etienne Roquain
}
