% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-functions.R
\name{delete.gaps}
\alias{delete.gaps}
\title{Delete the gaps induced by pruning}
\usage{
delete.gaps(C, ZL, leaf_list)
}
\arguments{
\item{C}{A list of list representing the forest structure. See \code{\link[=V.star]{V.star()}} for more information.}

\item{ZL}{A list of integer vectors representing the upper bounds \eqn{\zeta_k} of the forest structure. See \code{\link[=V.star]{V.star()}} for more information.}

\item{leaf_list}{A list of vectors representing the atoms of the forest structure. See \code{\link[=V.star]{V.star()}} for more information.}
}
\value{
A list with two named elements. \describe{
\item{\code{C}}{The new \code{C} after deleting the gaps.}
\item{\code{ZL}}{The new \code{ZL} after deleting the gaps.}
}
}
\description{
A small optimization that can be done after pruning
that can speed up computations (it removes the gaps introduced by the pruning)
}
\details{
See \code{\link[=pruning]{pruning()}}. This code has not been reviewed since 2023-07-25, use with caution.
}
