% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{dyadic.from.height}
\alias{dyadic.from.height}
\title{Dyadic tree structure from height}
\usage{
dyadic.from.height(m, H, method)
}
\arguments{
\item{m}{An integer value, the number of elements in the structure}

\item{H}{An integer value, the desired maximal height of the tree}

\item{method}{A numeric value. If \code{method == 1}, start from the leaves
and group nodes of a same height 2 by 2 as long as possible. If
\code{method==2}, start from the root and divide nodes in 2 nodes of equal
size as long as possible}
}
\description{
Dyadic tree structure from height
}
\examples{

m <- 6
dd <- dyadic.from.height(m, H = 3, method = 2)
str(dd)

}
