% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{dyadic.from.leaf_list}
\alias{dyadic.from.leaf_list}
\title{Create a complete dyadic tree structure}
\usage{
dyadic.from.leaf_list(leaf_list, method)
}
\arguments{
\item{leaf_list}{A list of leaves}

\item{method}{A numeric value. If \code{method == 1}, start from the leaves
and group nodes of a same height 2 by 2 as long as possible. If
\code{method==2}, start from the root and divide nodes in 2 nodes of equal
size as long as possible}
}
\value{
A list of lists containing the dyadic structure
}
\description{
Create a complete dyadic tree structure
}
