% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{dyadic.from.window.size}
\alias{dyadic.from.window.size}
\title{Dyadic tree structure from window size}
\usage{
dyadic.from.window.size(m, s, method)
}
\arguments{
\item{m}{An integer value, the number of elements in the structure}

\item{s}{An integer value, the number of elements in each leaf}

\item{method}{A numeric value. If \code{method == 1}, start from the leaves
and group nodes of a same height 2 by 2 as long as possible. If
\code{method==2}, start from the root and divide nodes in 2 nodes of equal
size as long as possible}
}
\value{
A list with two elements:\describe{
\item{leaf_list}{A list of leaves}
\item{C}{A list of lists containing the dyadic structure}
}
}
\description{
Dyadic tree structure from window size
}
\examples{

m <- 13
dd <- dyadic.from.window.size(m, s = 2, method = 2)
str(dd)


}
