% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{dyadic.max.height}
\alias{dyadic.max.height}
\title{Dyadic tree structure from maximum achievable height}
\usage{
dyadic.max.height(m, method)
}
\arguments{
\item{m}{An integer value, the number of elements in the structure}

\item{method}{A numeric value. If \code{method == 1}, start from the leaves
and group nodes of a same height 2 by 2 as long as possible. If
\code{method==2}, start from the root and divide nodes in 2 nodes of equal
size as long as possible}
}
\description{
Dyadic tree structure from maximum achievable height
}
\examples{

m<-1000
dd<-dyadic.from.max.height(m,s=10,method=2)
leaf_list<-dd$leaf_list

}
