% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Em.R
\name{Em}
\alias{Em}
\title{Title}
\usage{
Em(
  m,
  A,
  Pi,
  f0x,
  f1x,
  x,
  eps = 1e-04,
  maxit = 1000,
  h = 0.3,
  f0_known,
  approx = TRUE
)
}
\arguments{
\item{m}{}

\item{A}{}

\item{Pi}{}

\item{f0x}{}

\item{f1x}{}

\item{x}{}

\item{eps}{}

\item{maxit}{}

\item{h}{}

\item{f0_known}{}

\item{approx}{}
}
\value{

}
\description{
Title
}
