% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_tot_r.R
\name{Em_tot_approx}
\alias{Em_tot_approx}
\title{Title}
\usage{
Em_tot_approx(m, A, Pi, f0x, f1x, x, eps, maxit, h)
}
\arguments{
\item{m}{}

\item{A}{}

\item{Pi}{}

\item{f0x}{}

\item{f1x}{}

\item{x}{}

\item{eps}{}

\item{maxit}{}

\item{h}{}
}
\value{

}
\description{
Title
}
\examples{
m <- 200
A <- matrix(c(0.95, 0.05, 0.2, 0.80), 2, 2, byrow = T)
Pi <- c(0.8,0.2)
theta <- sim_markov(m, Pi , A)
x <- rep(0, m)
x[theta == 0] <- rnorm(sum(theta ==0))
x[theta == 1] <- rnorm(sum(theta ==1), 2, 1)
f0x <- dnorm(x)
f1x <- dnorm(x, 2,1)
Em_tot_approx(m, A, Pi, f0x, f1x,
 x, eps= 0.0001, maxit=1000, h= 0.3)
}
