% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empiricalCoverage.R
\name{empiricalCoverage}
\alias{empiricalCoverage}
\title{Empirical coverage of a threshold family}
\usage{
empiricalCoverage(thr, mat)
}
\arguments{
\item{thr}{A numeric vector of length \code{m}, a threshold family. Should  be sorted in decreasing order.}

\item{mat}{A \eqn{c} x \eqn{B} matrix of \code{B} Monte-Carlo
samples of \code{c} test statistics under the null hypothesis.}
}
\value{
The empirical coverage of \code{thr} with respect to \code{mat}.
}
\description{
Empirical coverage of a threshold family
}

