% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{empiricalCoverageO}
\alias{empiricalCoverageO}
\title{Empirical coverage of a threshold family}
\usage{
empiricalCoverageO(thr, Z)
}
\arguments{
\item{thr}{A numeric vector of length \code{m}, a threshold family. Should
be sorted in decreasing order.}

\item{Z}{A \code{m} x \code{B} matrix of \code{B} Monte-Carlo
samples of *ordered* test statistics under the null
hypothesis. Each column should be sorted in decreasing order.}
}
\value{
The empirical coverage of \code{thr} according to \code{Z}.
}
\description{
Empirical coverage of a threshold family
}
\details{
The empirical coverage of \code{thr} according to
  \code{Z} is defined as the empirical probability that there exists \eqn{k}
  in \eqn{1 \dots m} such that \eqn{k}-th max of the joint distribution of
  the test statistics under the null hypothesis exceeds the \eqn{k}-th
  largest value of \code{thr}, that is, \code{thr[k]}.
}
