% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{estimate_jer}
\alias{estimate_jer}
\title{Get a vector of pivotal statistics associated
to permutation p-values and to a reference family}
\usage{
estimate_jer(template, pval0, k_max)
}
\arguments{
\item{template}{matrix (B,m), template used for the calibration}

\item{pval0}{matrix (B,m) of permuted pvalues}

\item{k_max}{An integer value in \verb{[1,m]}, the number of elements in the reference family.}
}
\value{
A vector of length \code{B} pivotal statitics, whose j-th entry
corresponds to \code{psi(g_j.X)} with notation of the AoS 2020 paper cited
below (section 4.5)
}
\description{
Get a vector of pivotal statistics associated
to permutation p-values and to a reference family
}
\references{
Blanchard, G., Neuvial, P., & Roquain, E. (2020). Post hoc
confidence bounds on false positives using reference families. \emph{Annals of
Statistics}, 48(3), 1281-1303.
}
