% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_est.R
\name{Estimation}
\alias{Estimation}
\title{Title}
\usage{
Estimation(
  x,
  h = 0.3,
  m0_init,
  sd0_init,
  df_init,
  norm_init,
  max_pi0 = 0.99999,
  f0_known = TRUE,
  f0x_est = NULL,
  pval = NULL,
  plot = FALSE,
  size_plot = min(10000, length(x))
)
}
\arguments{
\item{x}{}

\item{h}{}

\item{m0_init}{}

\item{sd0_init}{}

\item{df_init}{}

\item{norm_init}{}

\item{max_pi0}{}

\item{f0_known}{}

\item{f0x_est}{}

\item{pval}{}

\item{plot}{}

\item{size_plot}{}
}
\value{

}
\description{
Title
}
\examples{
m <- 2000
theta <- sim_markov(m = 200 ,Pi = c(0.8,0.2), A = matrix(c(0.95, 0.05, 0.2, 0.80), 2, 2, byrow = T))
x <- rep(0, m)
 x[theta == 0] <- rnorm(sum(theta ==0))
 x[theta == 1] <- rnorm(sum(theta ==1), 2, 1)
 Estimation(x, m0_init = 0, sd0_init = 1, norm_init = TRUE, plot= TRUE) 
}
