% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_hat.R, R/long_reg.R
\name{f_hatK}
\alias{f_hatK}
\title{Non parametric density estimation}
\usage{
f_hatK(x_obs, x, h = 0.3, K)

f_hatK(x_obs, x, h = 0.3, K)
}
\arguments{
\item{x_obs}{}

\item{x}{}

\item{h}{}

\item{K}{}
}
\value{



}
\description{
Non parametric density estimation

Title
}
