% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-functions.R
\name{forest.completion}
\alias{forest.completion}
\title{Complete a forest structure}
\usage{
forest.completion(C, ZL, leaf_list)
}
\arguments{
\item{C}{A list of list representing the forest structure. See \code{\link[=V.star]{V.star()}} for more information.}

\item{ZL}{A list of integer vectors representing the upper bounds \eqn{\zeta_k} of the forest structure. See \code{\link[=V.star]{V.star()}} for more information.}

\item{leaf_list}{A list of vectors representing the atoms of the forest structure. See \code{\link[=V.star]{V.star()}} for more information.}
}
\value{
A list with two named elements. \describe{
\item{\code{C}}{The new \code{C} after completion.}
\item{\code{ZL}}{The new \code{ZL} after completion.}
}
}
\description{
Completes the forest in the sens of the Reference: adds the missing atoms/leafs
in the reference family with a forest structure \eqn{(R_k, \zeta_k)}
so that each atom is well represented by a \eqn{R_k}. The associated \eqn{\zeta_k} is
taken as the trivial \eqn{|R_k|}.
}
\details{
The forest must not be pruned (with \code{\link[=pruning]{pruning()}}) beforehand. The code will not behave expectedly
and will return a wrong result if a pruned forest is given as input. Maybe the function could be rewritten
going from the leaves to the roots instead of the contrary, to avoid this issue.
}
\references{
Durand, G., Blanchard, G., Neuvial, P., & Roquain, E. (2020). Post hoc false positive control for structured hypotheses. Scandinavian Journal of Statistics, 47(4), 1114-1148.

Durand G. (2025). A fast algorithm to compute a curve of confidence upper bounds for the False Discovery Proportion using a reference family with a forest structure. arXiv:2502.03849.
}
