% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc-bounds.R
\name{formatBounds}
\alias{formatBounds}
\title{Table of bounds}
\usage{
formatBounds(
  max_FP,
  idxs = seq_len(max_FP),
  lab = NULL,
  what = c("TP", "FDP"),
  all = FALSE
)
}
\arguments{
\item{max_FP}{a vector of the upper boud for the number of False Discovery}

\item{idxs}{vector of indices of the selection}

\item{lab}{label of the method used}

\item{what}{type of quantities. Should be in c("FP", "TP", "FDP", "TDP").}

\item{all}{Boolean to specify if the function return bounds for all tests}
}
\value{
data frame with the post hoc bounds
}
\description{
Table of bounds
}
