% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP.R
\name{FP}
\alias{FP}
\title{Upper bound for the number of false discoveries in a selection}
\usage{
FP(stat, thr)
}
\arguments{
\item{stat}{A vector of test statistics of the selected items}

\item{thr}{A vector of non-increasing \eqn{K} JER-controlling thresholds}
}
\value{
An upper bound on the number of false discoveries in the selection
}
\description{
Upper bound for the number of false discoveries in a selection
}
\examples{

m <- 200
B <- 1e3
rho <- 0.3
testStat <- gaussianTestStatistics(m, B, pi0 = 0.5, SNR = 2, dep = "equi", param = rho)
mat <- testStat$X0
stat <- testStat$x
cal <- jointFWERControl(mat = mat, refFamily = "Simes", alpha = 0.2)
cal$lambda  ## > alpha if rho > 0

thr <- sort(cal$thr, decreasing = TRUE)

M0 <- FP(stat, thr) ## upper bound on m0...
M0/m

sstat <- sort(stat, decreasing=TRUE)
FP(head(sstat), thr)
FP(c(head(sstat), tail(sstat)), thr)

}
