% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-simulation-functions.R
\name{gauss_bloc}
\alias{gauss_bloc}
\title{Generate a block of signal of Gaussian shape}
\usage{
gauss_bloc(barmu, len)
}
\arguments{
\item{barmu}{A numeric value, the strength of the signal}

\item{len}{An integer value, the length of the block}
}
\value{
A numeric vector of length \code{len}, the signal values in the block
}
\description{
Generate a block of signal of Gaussian shape
}
