% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-simulation-functions.R
\name{gen.mu.noleaves}
\alias{gen.mu.noleaves}
\title{Generate an unstructured signal}
\usage{
gen.mu.noleaves(m, pi0, barmu)
}
\arguments{
\item{m}{An integer value, the number of hypotheses}

\item{pi0}{A numeric value in \eqn{[0,1]}, the proportion of true null signals}

\item{barmu}{A numeric value, the strength of the signal}
}
\value{
A numeric vector of length \code{m}, the signal values for each hypothesis
}
\description{
Generate an unstructured signal
}
\examples{
m <- 100
pi0 <- 0.7
barmu <- 4
mu <- gen.mu.noleaves(m = m, pi0 = pi0, barmu = barmu) 
plot(mu)
}
