% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-simulation-functions.R
\name{gen.p.values}
\alias{gen.p.values}
\title{Generate one-sided p-values associated to a given signal with equi-correlated noise}
\usage{
gen.p.values(
  m,
  mu = rep(0, m),
  rho = 0,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{m}{An integer value, the number of hypotheses}

\item{mu}{A vector of \eqn{m} signal values}

\item{rho}{A numeric value in \eqn{[0,1]}, the level of equi-correlation between test statistics}
}
\value{
A vector of \eqn{m} one-sided \eqn{p}-values
}
\description{
Generate one-sided p-values associated to a given signal with equi-correlated noise
}
\examples{
m <- 100
s <- 10
K1 <- floor(m/(s * 4))
d <- 1
barmu <- 4
dd <- dyadic.from.window.size(m, s, method = 2)
leaf_list <- dd$leaf_list
mu <- gen.mu.leaves(m = m, K1 = K1, d = d, grouped = FALSE, 
                    setting = "const", barmu = barmu, leaf_list =leaf_list)
pvals <- gen.p.values(m = m, mu = mu, rho = 0)
plot(-log(pvals), t = 'b')
}
