% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{get_data_driven_template}
\alias{get_data_driven_template}
\title{For a given risk level, calibrate the method on learned templates by
dichotomy. This is equivalent to calibrating using pivotal stats but does
not require the availability of a closed form inverse template.}
\usage{
get_data_driven_template(
  X,
  label,
  B,
  row_test_fun = rowWelchTests,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{X}{A matrix (m, n)}

\item{label}{A numeric vector of \eqn{n} values in \eqn{0, 1}, the groups of
observations on which to perform two-sample tests}

\item{B}{integer, the number of permutation.}

\item{row_test_fun}{A (vectorized) test function. Defaults to \link{rowWelchTests}.}

\item{alternative}{A character string specifying the alternative hypothesis, to be passed to \code{rowTestFUN}. Must be one of "two.sided" (default), "greater" or "less".}
}
\value{
A matrix (m, B) of a learned template
}
\description{
For a given risk level, calibrate the method on learned templates by
dichotomy. This is equivalent to calibrating using pivotal stats but does
not require the availability of a closed form inverse template.
}
\examples{

m <- 50
n <- 45
X <- matrix(rnorm(m * n), ncol = n, nrow = m)
categ <- rbinom(n, 1, 0.4)
B <- 10
null_groups <- replicate(B, sample(categ))
p0 <- rowWelchTests(X, null_groups)$p.value
pivStat <- get_pivotal_stat(p0, m)
quantile(pivStat, 0.2)

}
\references{
Blanchard, G., Neuvial, P., & Roquain, E. (2020). Post hoc
confidence bounds on false positives using reference families. \emph{Annals of
Statistics}, 48(3), 1281-1303.
}
