% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{get_perm}
\alias{get_perm}
\title{Get permutation statistics and p-values}
\usage{
get_perm(
  X,
  categ,
  B,
  rowTestFUN = rowWelchTests,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{X}{A matrix of \code{m} variables (hypotheses) by \code{n} observations}

\item{categ}{An numeric vector of \code{n} values in \verb{0,1}
specifying the column indices of the first and second samples.}

\item{B}{An integer value, the number of permutations to be performed}

\item{rowTestFUN}{a vectorized testing function (same I/O as \link{rowWelchTests})}

\item{alternative}{A character string specifying the alternative hypothesis, to be passed to \code{rowTestFUN}. Must be one of "two.sided" (default), "greater" or "less".}
}
\description{
Get a matrix of statistics and p-values under the null hypothesis obtained by repeated permutation of class labels (two-sample tests)
}
\details{
The element 'p.value' of the output is a \verb{m x B} matrix whose entry i,j corresponds to \verb{p_\{(i)\}(g_j.X)} with notation of the AoS 2020 paper cited below (section 4.5).
}
\examples{
m <- 50
n <- 45
X <- matrix(rnorm(m*n), ncol = n, nrow = m)
categ <- rbinom(n, 1, 0.4)

B <- 10
set.seed(123)
perm0 <- sanssouci:::get_perm(X, categ, B, rowWelchTests)

# for this particular test 'get_perm' can be bypassed
set.seed(123)
null_groups <- replicate(B, sample(categ))
perm <- rowWelchTests(X, null_groups)
identical(perm0$p.value, perm$p.value)
identical(perm0$statistic, perm$statistic)
# Wilcoxon tests
set.seed(123)
perm0 <- sanssouci:::get_perm(X, categ, B, rowWilcoxonTests)
perm <- rowWilcoxonTests(X, null_groups)
identical(perm0$p.value, perm$p.value)
identical(perm0$statistic, perm$statistic)
}
