% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{get_randomized_p_values_one_sample}
\alias{get_randomized_p_values_one_sample}
\title{Get permutation  p-values}
\usage{
get_randomized_p_values_one_sample(
  X,
  B = 100,
  rowTestFUN = rowTtestsOneSample,
  seed = NULL
)
}
\arguments{
\item{X}{A matrix of \code{m} variables (hypotheses) by \code{n} observations}

\item{B}{An integer value, the number of permutations to be performed}

\item{rowTestFUN}{a vectorized testing function (same I/O as \link{rowWelchTests})}

\item{seed}{a single value, interpreted as an integer or NULL. The seed used for the Random Number Generation. NULL by default.}
}
\description{
Get a matrix of p-values under the null hypothesis obtained by sign-flipping (one-sample test).
}
\details{
The element 'p.value' of the output is a \verb{m x B} matrix whose entry i,j corresponds to \verb{p_\{(i)\}(g_j.X)} with notation of the AoS 2020 paper cited below (section 4.5).
}
