% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_borne.R
\name{get_zeta_ks}
\alias{get_zeta_ks}
\title{Title}
\usage{
get_zeta_ks(R_ks, al, for_back, Pis, f0x, f1x)
}
\arguments{
\item{R_ks}{list containing the position for the different Rk}

\item{al}{confidence level alpha}

\item{for_back}{result from the forward backward algorithm}

\item{f0x}{the values of the density under the null hypothesis for the different points}

\item{f1x}{the values of the density under the alternative hypothesis for the different points}

\item{L}{the L matrix (L[i,j] is the probability of theta_j = 0 knowing theta_i = 0 and X)}
}
\value{

}
\description{
Title
}
\examples{
 m <-  100
 A <- matrix(c(0.95, 0.05, 0.2, 0.80), 2, 2, byrow = T)
 f0 <- c(0, 1)
 f1 <- c(2, 1)
 Pi <- c( 0.9, 0.1)
 rdata <- simulate.data.hmm.2states(m, Pi, A, f0, f1)
 x <- rdata$x
 theta <- rdata$theta
 mod <- for_back(m, A, f0x, f1x, Pi)
viter_log <- viterbi_log(m,log(A),log(f0x),log(f1x), log(Pi))
Rk_s <- get_Rks(viterbi_log, min_size =3)
L <- get_L1(A, m,fw_bc_oracle$alpha, fw_bc_oracle$beta, f0x,  f1x)
Zeta <- get_zeta_ks(Rk_s, al = 0.2, for_back = fw_bc_oracle, L =L, f0x,  f1x)
get_phborne(S = 1:m, Rk_s, Zeta)
}
