% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFactorModelCovarianceMatrix.R
\name{getFactorModelCovarianceMatrix}
\alias{getFactorModelCovarianceMatrix}
\title{getFactorModelCovarianceMatrix}
\usage{
getFactorModelCovarianceMatrix(m, h = numeric(0), P = Matrix(nrow = m, ncol
  = length(h)), rho = 0)
}
\arguments{
\item{m}{Number of tests}

\item{h}{A vector of \code{k} singular values associated to each factor}

\item{P}{A \code{m x k} matrix of factor loadings}

\item{rho}{\code{1-rho} is the standard deviation of the noise}
}
\description{
calculate the covariance matrix of a factor model
}
\author{
Gilles Blanchard, Pierre Neuvial and Etienne Roquain
}
