% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowWelchTests.R
\name{getSummaryStats}
\alias{getSummaryStats}
\title{Convert a matrix of observations into summary statistics}
\usage{
getSummaryStats(mat, categ)
}
\arguments{
\item{mat}{A \code{m x n} numeric matrix whose rows correspond to variables
and columns to observations}

\item{categ}{A vector of \code{n} categories for the observations}
}
\value{
A list of \code{n} elements containing the above-described
summary statistics for each category
}
\description{
Convert a matrix of observations labelled into categories into summary
statistics for each category
}
\details{
The following statistics are calculated: sums, sums of squares, means,
standard deviations, sample sizes
}
\examples{

\dontrun{
# not exported
mat <- matrix(rnorm(3051*38), ncol = 38)
cls <- rep(c(0, 1), times = c(27, 11))
stats <- getSummaryStats(mat, categ = cls) 
}
}
\author{
Pierre Neuvial
}
