% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grzaphes.R
\name{graphes}
\alias{graphes}
\title{Title}
\usage{
graphes(nom, x, pos, Est_cnv, Sel, Final)
}
\arguments{
\item{nom}{}

\item{x}{}

\item{pos}{}

\item{Est_cnv}{}

\item{Sel}{}
}
\value{

}
\description{
Title
}
