% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k.max.R
\name{k.max}
\alias{k.max}
\title{k max}
\usage{
k.max(x, k)
}
\arguments{
\item{x}{A numeric vector}

\item{k}{An integer vector of test statistics of length m}
}
\value{
the k^th maximum of x
}
\description{
k^th maximum of a numeric vector
}
\examples{

x <- rnorm(100)
k.max(x, 10);
sort(x)[100-10+1]

}
\author{
Gilles Blanchard, Pierre Neuvial and Etienne Roquain
}

