% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kFWERPivotalStatistic.R
\name{kFWERPivotalStatistic}
\alias{kFWERPivotalStatistic}
\title{Pivotal statistic for the balanced threshold family of BNR}
\usage{
kFWERPivotalStatistic(mat, kMax = nrow(mat), C = 1:nrow(mat))
}
\arguments{
\item{mat}{A \eqn{c} x \eqn{B} matrix of \code{B} Monte-Carlo
samples of \code{c} test statistics under the null hypothesis,
with each sample (column) sorted decreasingly.}

\item{kMax}{For simultaneous control of (\eqn{k}-FWER for all \eqn{k \leq
k[max]}).}

\item{C}{A subset of \code{1:c} on which calibration is performed.}
}
\value{
the pivotal statistic \eqn{s_k} for the balanced threshold family introduced by BNR
}
\description{
Pivotal statistic for the balanced threshold family of BNR
}
