% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kFWERThresholdFamily.R
\name{kFWERThresholdFamily}
\alias{kFWERThresholdFamily}
\title{kFWERThresholdFamily}
\usage{
kFWERThresholdFamily(mat, kMax = nrow(mat), Rcpp = TRUE, verbose = FALSE)
}
\arguments{
\item{mat}{A \eqn{m} x \eqn{B} matrix of Monte-Carlo samples of test
statistics under the null hypothesis. \describe{ \item{m}{is the number of
tested hypotheses} \item{B}{is the number of Monte-Carlo samples}}}

\item{kMax}{For simultaneous control of (\eqn{k}-FWER for all \eqn{k \leq
k[max]}).}

\item{Rcpp}{If \code{TRUE} (the default), some costly operations (sorting)
are performed in C++.}

\item{verbose}{If \code{TRUE}, print results of intermediate calculations.
Defaults to \code{FALSE}.}
}
\value{
A function \eqn{f:alpha \mapsto f(alpha)} such that for each
\eqn{alpha}, \eqn{f(alpha)} is a candidate threshold family for controlling
the joint FWER.
}
\description{
calibration of the 'balanced' threshold family of BNR
}
\details{
This family \eqn{(t_k)} of thresholds is calibrated in such a way that for
each \eqn{k\leq kMax}, \eqn{(t_k)} controls the (marginal) k-FWER

The \eqn{m} x \eqn{B} matrix resulting of sorting the argument \code{mat}
by row and then by columns is returned as the attribute 'Q' of the return
value.
}
\examples{

m <- 1023
B <- 1e3

flavor <- c("independent", "equi-correlated", "3-factor model")[2]
rho <- 0.2
mat <- simulateGaussianNullsFromFactorModel(m, B, flavor=flavor, rho=rho)

sk <- kFWERThresholdFamily(mat)

thr <- sk(0.2)

}
\author{
Gilles Blanchard, Pierre Neuvial and Etienne Roquain
}

