% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapCalibration.R
\name{lm_test}
\alias{lm_test}
\title{Perform statistical test from contrast matrix for Linear model.}
\usage{
lm_test(Y, X, C)
}
\arguments{
\item{Y}{Expression matrix of size $n$ observations (in row) and $D$
dimensions/mesures in columns}

\item{X}{Design matrix of size $n$ observations (in row) and $p$ variables
(in columns)}

\item{C}{Contrast matrix of size $L$ tested contrast (in row) and $p$ columns
corresponding to te tested parameters.}
}
\value{
epsilon_est, the residuals matrix (size \eqn{n \times D}), stat_test the
test statistic matrix (size \eqn{L \times D}) and p-values, the p-values matrix
(size \eqn{L \times D})
}
\description{
Perform statistical test from contrast matrix for Linear model.
}
\examples{
N = 100
p = 2
D = 2
X <- matrix(0,nrow = N, ncol = p)
X[,1] <- 1
X[,-1] <- runif(N*(p-1), min = 0, max = 3)
beta <- matrix(0, nrow = p, ncol = D)
epsilons <- matrix(rnorm(N*D), nrow = N, ncol = D)
Y <- X \%*\% beta + epsilons
C <- diag(p)
resLM <- lm_test(Y = Y, X = X, C = C)
}
