% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{marginalKFWER}
\alias{marginalKFWER}
\title{Marginal Empirical coverage of a threshold family}
\usage{
marginalKFWER(thr, Z)
}
\arguments{
\item{thr}{A numeric vector of length \code{m}, a threshold family. Should
be sorted in decreasing order.}

\item{Z}{A \code{m} x \code{B} matrix of \code{B} Monte-Carlo samples of test statistics
under the null hypothesis. Each column should be sorted in decreasing order.}
}
\value{
The marginal empirical coverage of \code{thr} according to \code{Z}.
}
\description{
Marginal Empirical coverage of a threshold family
}
\details{
The marginal empirical coverage of \code{thr} according to
  \code{Z} is defined for an hypothesis of index \code{k} in \eqn{1 \dots m}
  as the empirical probability the \eqn{k}-th max of the joint distribution of
  the test statistics under the null hypothesis exceeds the \eqn{k}-th
  largest value of \code{thr}, that is, \code{thr[k]}.
}

