% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc-bounds.R
\name{maxFDP}
\alias{maxFDP}
\title{Upper bound for the false discovery proportion in a selection}
\usage{
maxFDP(p.values, thr)
}
\arguments{
\item{p.values}{A vector of p-values for the selected items}

\item{thr}{A vector of non-decreasing JER-controlling thresholds}
}
\value{
Upper bound on the proportion of false discoveries in the selection
}
\description{
Upper bound for the false discovery proportion in a selection
}
