% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc_bounds.R
\name{maxFP}
\alias{maxFP}
\title{Upper bound for the number of false discoveries in a selection}
\usage{
maxFP(stat, thr)
}
\arguments{
\item{stat}{A vector of test statistics of the selected items}

\item{thr}{A vector of non-increasing \eqn{K} JER-controlling thresholds}
}
\value{
An upper bound on the number of false discoveries in the selection
}
\description{
Upper bound for the number of false discoveries in a selection
}
\examples{

m <- 123
sim <- gaussianSamples(m = m, rho = 0.2, n = 100, 
                       pi0 = 0.8, SNR = 3, prob = 0.5)
X <- sim$X
cal <- calibrateJER(X, B = 1e3, alpha = 0.2, refFamily="Simes")
thr <- sort(cal$thr, decreasing = TRUE)
stat <- cal$stat

M0 <- maxFP(stat, thr) ## upper bound on m0...
M0/m

sstat <- sort(stat, decreasing=TRUE)
maxFP(head(sstat), thr)
maxFP(c(head(sstat), tail(sstat)), thr)

}
