% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{minPseudoRanks}
\alias{minPseudoRanks}
\title{Sorting of the rows of a matrix by descending order}
\usage{
minPseudoRanks(X, Y)
}
\arguments{
\item{X}{A numeric matrix.}

\item{Y}{A numeric matrix.}
}
\value{
A numeric matrix whose rows are sorted by descending order.
}
\description{
Sorting of the rows of a matrix by descending order
}

