% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc-bounds.R
\name{minTDP}
\alias{minTDP}
\title{Lower bound for the true discovery proportion in a selection}
\usage{
minTDP(p.values, thr)
}
\arguments{
\item{p.values}{A vector of p-values for the selected items}

\item{thr}{A vector of non-decreasing JER-controlling thresholds}
}
\value{
Lower bound on the proportion of true discoveries in the selection
}
\description{
Lower bound for the true discovery proportion in a selection
}
