% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP.R
\name{minTP}
\alias{minTP}
\title{Lower bound for the number of true discoveries in a selection}
\usage{
minTP(stat, thr)
}
\arguments{
\item{stat}{A vector of test statistics of the selected items}

\item{thr}{A vector of non-increasing \eqn{K} JER-controlling thresholds}
}
\value{
A Lower bound on the number of true discoveries in the selection
}
\description{
Lower bound for the number of true discoveries in a selection
}
