% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-functions.R
\name{nb.elements}
\alias{nb.elements}
\title{Number of unique regions in a reference family with forest structure}
\usage{
nb.elements(C)
}
\arguments{
\item{C}{A list of list representing the forest structure. See \code{\link[=V.star]{V.star()}} for more information.}
}
\value{
An integer, the number of regions.
}
\description{
Number of unique regions in a reference family with forest structure
}
