% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{partialColSortDescCpp}
\alias{partialColSortDescCpp}
\title{partial sorting of the columns of a matrix by descending order}
\usage{
partialColSortDescCpp(X, k)
}
\arguments{
\item{X}{A numeric matrix}

\item{k}{An integer value between 1 and \code{nrow(mat)}}
}
\description{
partial sorting of the columns of a matrix by descending order
}

