% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SansSouci-class.R
\name{plot.SansSouci}
\alias{plot.SansSouci}
\title{Plot confidence bound on the true/false positives among most significant items}
\usage{
\method{plot}{SansSouci}(x, y, xmax = nHyp(x), ...)
}
\arguments{
\item{x}{An object of class 'SansSouci'}

\item{y}{Not used}

\item{xmax}{Right limit of the plot}

\item{...}{Further arguments to be passed to \code{bound}}
}
\description{
Plot confidence bound on the true/false positives among most significant items
}
