% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceEnvelope.R
\name{plotConfidenceEnvelope}
\alias{plotConfidenceEnvelope}
\title{Plot confidence envelope}
\usage{
plotConfidenceEnvelope(conf_env, xmax, cols = NULL)
}
\arguments{
\item{conf_env}{A data.frame or a list of data.frames as output by 
\code{\link{confidenceEnvelope}}}

\item{xmax}{Right limit of the plot}

\item{cols}{A vector of colors of the same length as `conf_env`}
}
\description{
Plot confidence envelope
}
\examples{

# Generate Gaussian data and perform multiple tests
sim <- gaussianSamples(m = 502, rho = 0.3, n = 100, pi0 = 0.8, SNR = 3, prob = 0.5)
dat <- sim$X
rwt <- rowWelchTests(dat, categ=colnames(dat), alternative = "greater")

# calculate and plot confidence envelope
alpha <- 0.1
ce <- confidenceEnvelope(rwt$p.value, refFamily = "Simes", param = alpha)
plotConfidenceEnvelope(ce, xmax = 200) 

# calculate and plot several confidence envelopes
B <- 100
cal <- calibrateJER(X = dat, B = B, alpha = alpha, refFamily = "Simes")
cal_beta <- calibrateJER(X = dat, B = B, alpha = alpha, refFamily = "Beta", K = 20)
cec <- confidenceEnvelope(rwt$p.value, refFamily = "Simes", param = cal$lambda)
all_env <- list("Simes" = ce, 
                "Simes + calibration"= cal$conf_env, 
                "Beta + calibration" = cal_beta$conf_env)
plotConfidenceEnvelope(all_env, xmax = 200)

}
