% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pluralize}
\alias{pluralize}
\title{Match the singular and plural forms of a word with a quantity}
\usage{
pluralize(word, n)
}
\arguments{
\item{word}{a character, the word to adapt}

\item{n}{a numeric, the quantity}
}
\value{
a character
}
\description{
Match the singular and plural forms of a word with a quantity
}
