% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postHocBySimes.R
\name{posthocBySimes}
\alias{posthocBySimes}
\title{post hoc bound obtained from Simes' inequality}
\usage{
posthocBySimes(
  p,
  select,
  alpha,
  flavor = c("full step down", "one step down", "single step")
)
}
\arguments{
\item{p}{A numeric vector of \code{m} p-values for all tested hypotheses.}

\item{select}{A vector of indices in \eqn{[1, \dots m]} of the hypotheses to be selected.}

\item{alpha}{A numeric value, the significance level of the test procedure.}

\item{flavor}{A character value to specify whether to run the "single step",
"one step down" or "full step down" versions of the Simes posthoc bound.
Defaults to "full step down" (equivalent to cherry::pickSimes)}
}
\value{
A integer value, Simes's lower bound on the number of
correct rejections within the selected hypotheses
}
\description{
Lower bound on the number of correct rejections using Simes' reference
family
}
\examples{
m <- 1e3
m1 <- 200
p <- 1-pnorm(c(rnorm(m1, mean=4), rnorm(m-m1, mean=0)))
R <- union(1:10, sample(m, 10))
alpha <- 0.10
if (require("cherry")) {
  hom <- hommelFast(p)
  pickSimes(hom, R, silent=TRUE, alpha = alpha)
}
posthocBySimes(p, R, alpha=alpha)

p <- c(0, 0.01, 0.02, 0.05, 0.1)
alpha <- 0.2
R <- 1:length(p)
posthocBySimes(p, R, alpha, flavor = "single step")
posthocBySimes(p, R, alpha, flavor = "one step down")
posthocBySimes(p, R, alpha, flavor = "full step down")
if (require("cherry")) {
  hom <- hommelFast(p)
  pickSimes(hom, R, silent=TRUE, alpha = alpha)
}
}
\references{
Blanchard, G., Neuvial, P., & Roquain, E. (2020). Post hoc confidence bounds on false positives using reference families. Annals of Statistics, 48(3), 1281-1303.

Goeman, J. J., & Solari, A. (2011). Multiple testing for exploratory research. Statistical Science, 26(4), 584-597.
}
\author{
Gilles Blanchard, Pierre Neuvial and Etienne Roquain
}
