% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SansSouci-class.R
\name{predict.SansSouci}
\alias{predict.SansSouci}
\title{Post hoc confidence bounds on the true/false positives}
\usage{
\method{predict}{SansSouci}(
  object,
  S = seq_len(nHyp(object)),
  what = c("TP", "FDP"),
  all = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class 'SansSouci'}

\item{S}{A subset of indices}

\item{what}{A character vector, the names of the post hoc bounds to be
computed, among:
\itemize{
\item FP: Upper bound on the number of false positives in the 'x' most significant items
\item TP: Lower bound on the number of true positives in the 'x' most significant items
\item FDP: Upper bound on the proportion of false positives in the 'x' most significant items
\item TP: Lower bound on the proportion of true positives in the 'x' most significant items
}

Defaults to \code{c("TP", "FDP")}}

\item{all}{A logical value: should the bounds for all ordered subsets of \code{S} be returned? If \code{FALSE} (the default), only the bound for \code{S} is returned}

\item{...}{Not used}
}
\value{
If \code{all} is \code{FALSE} (the default), only the value of the bound is returned. Otherwise, a \code{data.frame} is return, with |S| rows and 4 columns:
\itemize{
\item x: Number of most significant items selected
\item label: Label for the procedure, typically of the form 'family(param)'
\item bound: Value of the post hoc bound
\item stat: Type of post hoc bound, as specified by argument \code{bound}.
}
}
\description{
Post hoc confidence bounds on the true/false positives
}
\examples{

# Generate Gaussian data and perform multiple tests
obj <- SansSouciSim(m = 502, rho = 0.5, n = 100, pi0 = 0.8, SNR = 3, prob = 0.5)
res <- fit(obj, B = 100, alpha = 0.1)

# post hoc bound on the set of hypotheses
predict(res)

# idem for all possible subsets (sorted by p-value)
bounds <- predict(res, all=TRUE)
head(bounds)

# post hoc bound on a subset
S <- which(pValues(res) < 0.01)
predict(res, S)
}
