% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-functions.R
\name{pruning}
\alias{pruning}
\title{Prune a forest structure to speed up computations}
\usage{
pruning(C, ZL, leaf_list, prune.leafs = FALSE, delete.gaps = FALSE)
}
\arguments{
\item{C}{A list of list representing the forest structure. See \code{\link[=V.star]{V.star()}} for more information.}

\item{ZL}{A list of integer vectors representing the upper bounds \eqn{\zeta_k} of the forest structure. See \code{\link[=V.star]{V.star()}} for more information.}

\item{leaf_list}{A list of vectors representing the atoms of the forest structure. See \code{\link[=V.star]{V.star()}} for more information.}

\item{prune.leafs}{A boolean, \code{FALSE} by default. If \code{TRUE}, will also prune atoms/leafs for which \eqn{\zeta_k \geq |R_k|},
this makes the computation of \code{\link[=V.star]{V.star()}} and \code{\link[=curve.V.star.forest.naive]{curve.V.star.forest.naive()}} even faster but should not be used with \code{\link[=curve.V.star.forest.fast]{curve.V.star.forest.fast()}}
because this needs the structure to be complete (i.e., with all its atoms). This is why the default option is \code{FALSE}.}

\item{delete.gaps}{A boolean, \code{FALSE} by default. If \code{TRUE}, will also delete the gaps in the structure induced
by the pruning, see \code{\link[=delete.gaps]{delete.gaps()}}.}
}
\value{
A list with three named elements. \describe{
\item{\code{VstarNm}}{\eqn{V^*(\mathbb N_m)} is computed as by-product by the algorithm, so we might as well return it.}
\item{\code{C}}{The new \code{C} after pruning.}
\item{\code{ZL}}{The new \code{ZL} after pruning.}
}
}
\description{
The pruned forest structure makes the computation of \code{\link[=V.star]{V.star()}},
\code{\link[=curve.V.star.forest.naive]{curve.V.star.forest.naive()}} and \code{\link[=curve.V.star.forest.fast]{curve.V.star.forest.fast()}} faster.
}
\references{
Durand, G., Blanchard, G., Neuvial, P., & Roquain, E. (2020). Post hoc false positive control for structured hypotheses. Scandinavian Journal of Statistics, 47(4), 1114-1148.

Durand G., preprint to appear with the description of pruning
}
