% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference-families.R
\name{reference-families}
\alias{reference-families}
\alias{t_inv_linear}
\alias{t_linear}
\alias{t_inv_beta}
\alias{t_beta}
\title{Reference families}
\usage{
t_inv_linear(y, k, m)

t_linear(alpha, k, m)

t_inv_beta(y, k, m)

t_beta(alpha, k, m)
}
\arguments{
\item{y}{numeric value}

\item{k}{integer value in \verb{[1,m]}}

\item{m}{integer value}

\item{alpha}{numeric value in (0,1)}
}
\description{
Reference families
}
\examples{
m <- 10
alpha <- 0.05
thr <- t_linear(alpha, 1:m, m)
t_inv_linear(thr[3], 3, m)

thr <- t_beta(alpha, 1:m, m)
t_inv_beta(thr[3], 3, m)
}
