% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapCalibration.R
\name{row_lm_test}
\alias{row_lm_test}
\title{Mass-univariate bootstrap-based inference for contrasts in a linear model}
\usage{
row_lm_test(
  Y,
  X,
  C,
  alternative = c("two.sided", "less", "greater"),
  groups = matrix(1:ncol(Y), ncol = 1)
)
}
\arguments{
\item{Y}{A data matrix of size $n$ observations (in row) and $D$ features in
columns}

\item{X}{A design matrix of size $n$ observations (in row) and $p$ variables
(in columns)}

\item{C}{A contrast matrix of size $L$ tested contrasts (in row) and $p$
columns corresponding to the parameters to be tested}

\item{alternative}{A character string specifying the alternative hypothesis.
Must be one of "two.sided" (default), "greater" or "less".}

\item{groups}{A numeric matrix of \eqn{n} rows and \eqn{B} columns values in
\eqn{1, ..., n},  the indicator
of the sample used in the test.}
}
\value{
A list containing the following components:
\describe{
\item{statistic}{the value of the t-statistics}
\item{p.value}{the p-values for the tests}
Each of these elements is a matrix of size \code{m x B},
coerced to a vector of length \code{m} if \code{B=1}
}
}
\description{
Compute the marginal null t-statistics for a set of contrasts and their
(two-sided) p-value by bootstrapping the residuals
}
