% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowPearsonCorrelationTests.R
\name{rowPearsonCorrelationTests}
\alias{rowPearsonCorrelationTests}
\title{Pearson's correlation test for rows of a matrix}
\usage{
rowPearsonCorrelationTests(X, categ, alternative = "two.sided")
}
\arguments{
\item{X}{A \code{m x n} numeric matrix whose rows correspond to variables
and columns to observations}

\item{categ}{Either a numeric vector of continuous covariate for
the observations}

\item{alternative}{A character string specifying the alternative hypothesis.
Must be one of "two.sided" (default), "greater" or "less". As in
\code{\link{t.test}}, alternative = "greater" is the alternative that class
1 has a larger mean than class 0.}
}
\value{
A list containing the following components:
\describe{
\item{statistic}{the value of the t-statistics}
\item{parameter}{the degrees of freedom for the t-statistics}
\item{p.value}{the p-values for the tests}
\item{estimate}{the correlation}}
Each of these elements is a matrix of size \code{m x B}, coerced to a vector of length \code{m} if \code{B=1}
}
\description{
Pearson's correlation test for rows of a matrix
}
\details{
This function is a wrapper around the \code{row_cor_pearson function} in the 'matrixTests' package.
}
\examples{

m <- 300
n <- 38
mat <- matrix(rnorm(m*n), ncol=n)
categ <- rnorm(n, mean = 10)
system.time(fwt <- rowPearsonCorrelationTests(mat, categ, alternative = "greater"))
str(fwt)

}
\seealso{
\code{\link[matrixTests:row_cor_pearson]{matrixTests::row_cor_pearson()}}
}
\author{
Nicolas Enjalbert Courrech
}
