% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowTtestsOneSample.R
\name{rowTtestsOneSample}
\alias{rowTtestsOneSample}
\title{One sample T tests for rows of a matrix}
\usage{
rowTtestsOneSample(
  mat,
  labels,
  alternative = c("two.sided", "grater", "less"),
  mu = 0
)
}
\arguments{
\item{mat}{A numeric matrix whose rows correspond to \code{m} variables and columns to \code{n}
observations}

\item{labels}{Either a numeric vector of \code{n} values in \eqn{-1, 1} to
perform sign flips, or a \code{n x B} matrix stacking \code{B} such
vectors. If missing, coerced to \code{rep(1, n)}}

\item{alternative}{A character string specifying the alternative hypothesis.
Must be one of "two.sided" (default), "greater" or "less". \code{Alternative =  "greater"} corresponds to a positive mean.}

\item{mu}{A number indicating the true value of the mean (or difference in
means.}
}
\value{
A list containing the following components:
\describe{
\item{statistic}{the value of the statistics}
\item{p.value}{the p-values for the tests}}
Each of these elements is a matrix of size \code{nrow(mat) x B}, coerced to a vector of length \code{nrow(mat)} if \code{B=1}
}
\description{
One sample T tests for rows of a matrix
}
\author{
Pierre Neuvial & Nicolas Enjalbert Courrech
}
