% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowZTests.R
\name{rowZTests}
\alias{rowZTests}
\title{Z tests for rows of a matrix}
\usage{
rowZTests(mat, categ, alternative)
}
\arguments{
\item{mat}{A numeric matrix whose rows correspond to \code{m} variables and columns to \code{n}
observations}

\item{categ}{Either a numeric vector of \code{n} values in \eqn{-1, 1} to
perform sign flips, or a \code{n x B} matrix stacking \code{B} such
vectors. If missing, coerced to \code{rep(1, n)}}

\item{alternative}{A character string specifying the alternative hypothesis.
Must be one of "two.sided" (default), "greater" or "less". \code{Alternative =  "greater"} corresponds to a positive mean.}
}
\value{
A list containing the following components:
\describe{
\item{statistic}{the value of the statistics}
\item{p.value}{the p-values for the tests}}
Each of these elements is a matrix of size \code{nrow(mat) x B}, coerced to a vector of length \code{nrow(mat)} if \code{B=1}
}
\description{
Z tests for rows of a matrix
}
\examples{

p <- 1e4+1
n <- 380
mat <- matrix(rnorm(p*n, mean = 1), ncol=n)
zt <- rowZTests(mat, alternative = "greater")
str(zt)

# compare with apply version:
p <- apply(mat, 1, FUN=function(x) {
   stat <- sum(x)/sqrt(length(x))
   pnorm(stat, lower.tail = FALSE)
})
all(abs(zt$p.value - p) < 1e-10)  ## same results

# Sign flipping
B <- 10
eps <- replicate(B, rbinom(n, 1, 0.5)*2 - 1)  ## Rademacher
zt_perm <- rowZTests(mat, eps, alternative = "greater")
str(zt_perm)
}
\author{
Pierre Neuvial
}
