% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SansSouci-class.R
\name{SansSouci-class}
\alias{SansSouci-class}
\alias{nHyp.SansSouci}
\alias{nObs.SansSouci}
\alias{label.SansSouci}
\alias{print.SansSouci}
\alias{pValues.SansSouci}
\alias{foldChanges.SansSouci}
\alias{thresholds.SansSouci}
\title{SansSouci class}
\usage{
\method{nHyp}{SansSouci}(object)

\method{nObs}{SansSouci}(object)

\method{label}{SansSouci}(object)

\method{print}{SansSouci}(x, ..., verbose = FALSE)

\method{pValues}{SansSouci}(object)

\method{foldChanges}{SansSouci}(object)

\method{thresholds}{SansSouci}(object)
}
\arguments{
\item{object}{An object of class \code{SansSouci}}

\item{x}{An object of class `SansSouci`}

\item{...}{Not used}

\item{verbose}{Should detailed output be printed? Defaults to FALSE}
}
\description{
SansSouci class
}
\examples{
data(expr_ALL, package = "sansSouci.data")
groups <- ifelse(colnames(expr_ALL)=="NEG", 0, 1)
table(groups)
a <- SansSouci(Y = expr_ALL, groups = groups)
print(a)
nHyp(a)
nObs(a)
label(a)

res <- fit(a, B = 100, alpha = 0.1)
print(res)
label(res)
}
