% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SansSouci-class.R
\name{SansSouci-methods}
\alias{SansSouci-methods}
\alias{nHyp.SansSouci}
\alias{nObs.SansSouci}
\alias{label.SansSouci}
\alias{print.SansSouci}
\alias{pValues.SansSouci}
\alias{foldChanges.SansSouci}
\alias{thresholds.SansSouci}
\title{Basic methods for class \code{SansSouci}}
\usage{
\method{nHyp}{SansSouci}(object)

\method{nObs}{SansSouci}(object)

\method{label}{SansSouci}(object)

\method{print}{SansSouci}(x, ..., verbose = FALSE)

\method{pValues}{SansSouci}(object)

\method{foldChanges}{SansSouci}(object)

\method{thresholds}{SansSouci}(object)
}
\arguments{
\item{object}{An object of class \code{SansSouci}}

\item{x}{An object of class \code{SansSouci}}

\item{...}{Not used}

\item{verbose}{Should detailed output be printed? Defaults to FALSE}
}
\description{
Basic methods for class \code{SansSouci}

\code{nHyp}: get the number of hypotheses

\code{nObs} Get the number of observations

\code{label} Get the label of a post hoc method

Print 'SansSouci' objects

\code{pValues}: get p-values

\code{thresholds}: get thresholds
}
\examples{
data(expr_ALL, package = "sanssouci.data")
groups <- ifelse(colnames(expr_ALL)=="NEG", 0, 1)
table(groups)
a <- SansSouci(Y = expr_ALL, groups = groups)
print(a)
nHyp(a)
nObs(a)
label(a)

res <- fit(a, B = 100, alpha = 0.1)
label(res)
print(res)
str(pValues(res))
str(foldChanges(res)) 
str(thresholds(res))
volcanoPlot(res, q = 0.05, r = 0.5)

}
