% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SansSouciStruct-class.R
\name{SansSouciStruct-methods}
\alias{SansSouciStruct-methods}
\alias{nHyp.SansSouciStruct}
\alias{nLeaves.SansSouciStruct}
\alias{label.SansSouciStruct}
\alias{pValues.SansSouciStruct}
\title{Basic methods for class \code{SansSouciStruct}}
\usage{
\method{nHyp}{SansSouciStruct}(object)

\method{nLeaves}{SansSouciStruct}(object)

\method{label}{SansSouciStruct}(object)

\method{pValues}{SansSouciStruct}(object)
}
\arguments{
\item{object}{An object of class \code{SansSouci}}
}
\description{
Basic methods for class \code{SansSouciStruct}

\code{nHyp}: get the number of hypotheses

\code{nLeaves}: get the number of leaves

\code{label} Get the label of a post hoc method

\code{pValues}: get p-values
}
\examples{
s <- 100
q <- 7
m <- s*2^q

obj <- SansSouciDyadic(m, param = s, flavor = "window.size", direction = "top-down")
s <- 100
q <- 7
m <- s*2^q
obj <- SansSouciDyadic(m, s, flavor = "window.size", direction = "top-down")
nHyp(obj)
nLeaves(obj)
mu <- gen.mu.leaves(m = m, K1 = 8, d = 0.9, grouped = TRUE, 
  setting = "const", barmu = 3, leaf_list = obj$input$leaves)
pvalues <- gen.p.values(m = m, mu = mu, rho = 0)

alpha <- 0.05
res <- fit(obj, alpha, pvalues, "DKWM")
label(res)
str(pValues(res))

}
