% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SansSouciStruct-class.R
\name{SansSouciStruct}
\alias{SansSouciStruct}
\title{Create an object of class 'SansSouciStruct'}
\usage{
SansSouciStruct(structure, leaves, truth = NULL)
}
\arguments{
\item{structure}{A (dyadic) structure}

\item{leaves}{A list of leaves}

\item{truth}{An optional numeric vector of $m$ values in ${0,1}$, the status of each null hypothesis (0 means H0 is true, 1 means H1 is true). Typically used in simulations.}
}
\description{
Create an object of class 'SansSouciStruct'
}
\examples{
s <- 100
q <- 7
m <- s*2^q

dd <- dyadic.from.window.size(m, s, method = 2)
leaf_list <- dd$leaf_list
C <- dd$C
obj1 <- SansSouciStruct(C, leaf_list)
obj <- SansSouciDyadic(m, s, flavor = "window.size", direction = "top-down")
}
