% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{selection}
\alias{selection}
\title{Title}
\usage{
selection(type_sel, EM, stuff, seuil_lis, seuil_p, min_size = 0)
}
\arguments{
\item{type_sel}{}

\item{EM}{}

\item{stuff}{}

\item{min_size}{}
}
\value{

}
\description{
Title
}
