% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimesPivotalStatistic.R
\name{SimesPivotalStatistic}
\alias{SimesPivotalStatistic}
\title{Simes' pivotal statistic}
\usage{
SimesPivotalStatistic(mat, kMax, m)
}
\arguments{
\item{mat}{A \eqn{c} x \eqn{B} matrix of \code{B} Monte-Carlo
samples of \code{c} test statistics under the null hypothesis.}

\item{kMax}{For simultaneous control of (\eqn{k}-FWER for all \eqn{k \leq
k[max]}).}

\item{m}{An integer value, the total number of hypotheses
tested. \code{m} should not be less than \code{c}.}
}
\value{
the pivotal statistic \eqn{s_k} of BNR for Simes'
threshold family
}
\description{
The pivotal statistic for Simes' threshold family
}
\details{
The inverse function of \eqn{s_k} is: \eqn{s_k^{-1}:u
\mapsto (m/k)*(1-pnorm(u))}.
}

