% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_tot.R
\name{simu_tot}
\alias{simu_tot}
\title{Title}
\usage{
simu_tot(
  m,
  A,
  Pi,
  n,
  rho,
  SNR,
  prob,
  type_sim,
  al,
  s_dbnr,
  b_act,
  d,
  seuil,
  h = 0.3,
  n_boot,
  min_size,
  norm,
  m0,
  sd0,
  df,
  m0_init,
  sd0_init,
  df_init,
  norm_init,
  max_pi0 = 0.99999,
  type_init,
  num_seed,
  f0_known,
  approx
)
}
\arguments{
\item{m}{}

\item{A}{}

\item{Pi}{}

\item{n}{}

\item{rho}{}

\item{SNR}{}

\item{prob}{}

\item{type_sim}{}

\item{al}{}

\item{s_dbnr}{}

\item{b_act}{}

\item{d}{}

\item{seuil}{}

\item{h}{}

\item{n_boot}{}

\item{min_size}{}

\item{norm}{}

\item{m0}{}

\item{sd0}{}

\item{df}{}

\item{m0_init}{}

\item{sd0_init}{}

\item{df_init}{}

\item{norm_init}{}

\item{max_pi0}{}
}
\value{

}
\description{
Title
}
\examples{
simu_tot( m = c(100),
A = matrix(c(0.95, 0.05, 0.2, 0.80), 2, 2, byrow = T),
Pi = c(0.95, 0.05),
n = c(500),
rho = c(0),
SNR = 2,
prob = c(0.5),
type_sim = c("HMM"),
n_boot = 20,
al = 0.2, s_dbnr = 10, b_act = 2, d = 1, seuil= 0.05,
min_size = 2, norm = TRUE, sd0 = 0.5, m0= 0,sd0_init = 0.5, m0_init= 0, norm_init = TRUE, df= 2)
}
