% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateEquiByGroupPermutation.R
\name{simulateEquiByGroupPermutation}
\alias{simulateEquiByGroupPermutation}
\title{Simulate test statistics in the two-group model}
\usage{
simulateEquiByGroupPermutation(m, rho, n, B, pi0, SNR = 1, p = 0.5,
  w = NULL)
}
\arguments{
\item{m}{Number of hypotheses}

\item{rho}{Level of equi-correlation between pairs of variables}

\item{n}{Number of observations}

\item{B}{Number of resamplings to estimate the test statistics}

\item{pi0}{Proportion of true null hypotheses}

\item{SNR}{Signal to noise ratio. Either a numeric value (a measure of 
distance between H0 and H1) or a vector of length \code{m*(1-pi0)}}

\item{p}{Probability of success of the outcome variable}

\item{w}{An optional vector of length \code{n}, the underlying factor driving
equi-correlation}
}
\value{
A list with elements \describe{ \item{x}{A vector of length \eqn{m}
  test statistics} \item{X0}{An \eqn{m x B} matrix of test statistics under
  the null hypothesis} \item{H}{A vector of length \eqn{m}, the status of
  each hypothesis: 0 for true null hypothesis, and 1 for true alternative 
  hypothesis}} The test statistics are \eqn{\sim N(0,1)}, and \eqn{\sim
  N(\mu,1)}, with \eqn{\mu>0}
}
\description{
Simulate test statistics in the two-group model
}
\examples{

m <- 123
rho <- 0.2
n <- 100
pi0 <- 0.5
B <- 1e3

sim <- simulateEquiByGroupPermutation(m, rho, n, B, pi0, SNR=2)
scoreMat <- sim$X0
stat <- sim$x

## show test statistics
pch <- 20
colStat <- 1+sim$H
plot(stat, col=colStat, main="Test statistics", pch=pch)
legend("topleft", c("H0", "H1"), pch=pch, col=1:2)

}
\author{
Gilles Blanchard, Pierre Neuvial and Etienne Roquain
}
