% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/simulateFactorModelNullsFromSingularValuesAndLoadings.R
\name{simulateFactorModelNullsFromSingularValuesAndLoadings}
\alias{simulateFactorModelNullsFromSingularValuesAndLoadings}
\title{simulateFactorModelNullsFromSingularValuesAndLoadings}
\usage{
simulateFactorModelNullsFromSingularValuesAndLoadings(m, h = numeric(0),
  P = Matrix(nrow = m, ncol = length(h)), rho = 0)
}
\arguments{
\item{m}{Number of tests}

\item{h}{A vector of \code{k} singular values associated to each factor}

\item{P}{A \code{m x k} matrix of factor loadings}

\item{rho}{\code{1-rho} is the standard deviation of the noise}
}
\value{
\item{Y}{a vector of \code{m} simulated observations} \item{W}{a
vector of \code{k} factors}
}
\description{
Simulate null hypotheses according to a factor model
}
\examples{

m <- 10

## independent
sim0 <- simulateFactorModelNullsFromSingularValuesAndLoadings(m)
str(sim0)
sum(is.na(simulateFactorModelNullsFromSingularValuesAndLoadings(m)))
S0 <- getFactorModelCovarianceMatrix(m)
image(S0)

## equi-correlated
rho <- 0.2
h <- 1
P <- matrix(1, m, length(h))
sim1 <- simulateFactorModelNullsFromSingularValuesAndLoadings(m, h, P, rho=rho)
str(sim1)
S1 <- getFactorModelCovarianceMatrix(m, h, P, rho=rho)
image(S1)

## 3-factor model
m <- 4*floor(m/4) ## make sure m/4 is an integer
rho <- 0.5
h <- c(0.5, 0.3, 0.2)*m
gamma1 <- rep(1,m)
gamma2 <- rep(c(-1, 1), each=m/2)
gamma3 <- rep(c(-1, 1), times=m/2)
P <- cbind(gamma1, gamma2, gamma3)/sqrt(m)

sim3 <- simulateFactorModelNullsFromSingularValuesAndLoadings(m, h, P, rho=rho)
str(sim3)
S3 <- getFactorModelCovarianceMatrix(m, h, P, rho=rho)
image(S3)
sim3

}
\author{
Gilles Blanchard, Pierre Neuvial and Etienne Roquain
}
