% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGaussianEquiCorrelatedNulls.R
\name{simulateGaussianEquiCorrelatedNulls}
\alias{simulateGaussianEquiCorrelatedNulls}
\title{simulateGaussianEquiCorrelatedNulls}
\usage{
simulateGaussianEquiCorrelatedNulls(m, n = 1000, rho = 0, w = NULL)
}
\arguments{
\item{m}{Number of tests}

\item{n}{Number of replications}

\item{rho}{\code{1-rho} is the standard deviation of the noise}

\item{w}{An optional vector of length \code{n}, the underlying factor
driving equi-correlation}
}
\value{
A \code{m x n} matrix of simulated observations. The underlying
factor driving equi-correlation is given as \code{attr(Y, "w")}.
}
\description{
Simulate equi-correlated null hypotheses
}
\examples{

m <- 1331
n <- 111
## equi-correlated
rho <- 0.2

Y <- simulateGaussianEquiCorrelatedNulls(m, n, rho)
## check equi-correlation:
covmat <- cov(t(Y))
diag(covmat) <- NA
dim(covmat) <- NULL
summary(covmat)

}
\author{
Gilles Blanchard, Pierre Neuvial and Etienne Roquain
}

