% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGaussianNullsFromSigma.R
\name{simulateGaussianNullsFromSigma}
\alias{simulateGaussianNullsFromSigma}
\title{simulateGaussianNullsFromSigma}
\usage{
simulateGaussianNullsFromSigma(Sigma, n = 1)
}
\arguments{
\item{Sigma}{Population covariance matrix (of size \code{m x m})}

\item{n}{Number of replications of the simulation}
}
\value{
A \code{m x n} \code{Matrix} simulated test statistics, where
\code{m} is the dimension of \code{Sigma}
}
\description{
Simulate m observations of a p-dimensional Gaussian vector with given
covariance matrix
}
\examples{

library("Matrix")
m <- 100
n <- 1000

## Toeplitz, short range
tcoefs <- toeplitz((1:m)^(-2))
Sigma <- Matrix(tcoefs, sparse = TRUE)
Y <- sansSouci:::simulateGaussianNullsFromSigma(Sigma, n)
str(Y)

SigmaHat <- Y \%*\% t(Y)/n
image(SigmaHat)
image(Sigma)

svd(SigmaHat)$d
(ss <- svd(Sigma)$d)
max(ss)/min(ss)


## Toeplitz, long range
tcoefs <- toeplitz((1:m)^(-.1))
Sigma <- Matrix(tcoefs, sparse = TRUE)
Y <- sansSouci:::simulateGaussianNullsFromSigma(Sigma, n)
str(Y)

SigmaHat <- Y \%*\% t(Y)/n
image(SigmaHat)
image(Sigma)

svd(SigmaHat)$d
(sl <- svd(Sigma)$d)
max(sl)/min(sl)

plot(ss/sum(ss), col=1, t='b', ylim=c(0,1))
lines(sl/sum(sl), col=2, t='b')

}
\author{
Gilles Blanchard, Pierre Neuvial and Etienne Roquain
}
