% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateToeplitz.R
\name{simulateToeplitz}
\alias{simulateToeplitz}
\title{Simulate Gaussian test statistics with Toeplitz covariance}
\usage{
simulateToeplitz(m, pow, B, pi0, SNR = 1)
}
\arguments{
\item{m}{Number of hypotheses}

\item{pow}{Exponent in the Toeplitz vector}

\item{B}{Number of simulations}

\item{pi0}{Proportion of true null hypotheses}

\item{SNR}{Signal to noise ratio. Either a numeric value (a measure of
distance between H0 and H1) or a vector of length \code{m*(1-pi0)}}
}
\value{
A list with elements \describe{
\item{x}{A vector of length \eqn{m} test statistics}
\item{X0}{An \eqn{m x B} matrix of test statistics under the null
hypothesis}
\item{H}{A vector of length \eqn{m}, the status of each
hypothesis: 0 for true null hypothesis, and 1 for true alternative
hypothesis} }
}
\description{
Simulate Gaussian test statistics with Toeplitz covariance
}
\examples{

m <- 123
pow <- -0.5
B <- 100
pi0 <- 0.5

sim <- simulateToeplitz(m, pow, B, pi0, SNR=1)
stat <- sim$x

## show test statistics
pch <- 20
colStat <- 1+sim$H
plot(stat, col=colStat, main="Test statistics", pch=pch)
legend("topleft", c("H0", "H1"), pch=pch, col=1:2)

}
\author{
Gilles Blanchard, Pierre Neuvial and Etienne Roquain
}
