% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testBySignFlipping.R
\name{testBySignFlipping}
\alias{testBySignFlipping}
\title{Test by sign flipping in a location model}
\usage{
testBySignFlipping(X, B, p.value = TRUE, seed = NULL)
}
\arguments{
\item{X}{a matrix of \code{m} variables by \code{n} observations}

\item{B}{a numeric value, the number of flips to be performed}

\item{p.value}{a boolean value: should \eqn{p}-values be calculated and 
returned? Defaults to @TRUE}

\item{seed}{an integer (or NULL) value used as a seed for random number 
generation. If \code{NULL}, no seed is specified.}
}
\value{
a list with two or four elements: \describe{
\item{T}{A vector of \eqn{m} test statistics}
\item{T0}{A \eqn{m \times B} matrix of permutation test statistics}
\item{p}{A vector of \eqn{m} \eqn{p}-values (only if \code{p.value} is @TRUE)}
\item{p0}{A \eqn{m \times B} matrix of permutation \eqn{p}-values (ony if \code{p.value} is @TRUE)}
}
}
\description{
Test by sign flipping in a location model
}
\examples{

p <- 513
n <- 38
mat <- matrix(rnorm(p*n), ncol=n)
resFlip <- testBySignFlipping(X=mat, B=1000, seed=123)

alpha <- 0.05
res <- jointFWERControl(resFlip$T0, refFamily="Simes", alpha=alpha, stat=resFlip$T)

}
