% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testByTwoSamplePermutation.R
\name{testByTwoSamplePermutation}
\alias{testByTwoSamplePermutation}
\title{Perform permutation testing for equal means in two groups}
\usage{
testByTwoSamplePermutation(X, cls, B, p.value = TRUE, seed = NULL)
}
\arguments{
\item{X}{a matrix of \code{m} variables by \code{n} observations}

\item{cls}{a vector of length \code{n} class labels in \code{0,1}}

\item{B}{a numeric value, the number of permutations to be performed}

\item{p.value}{a boolean value: should \eqn{p}-values be calculated and 
returned? Defaults to @TRUE}

\item{seed}{an integer (or NULL) value used as a seed for random number 
generation. If \code{NULL}, no seed is specified.}
}
\value{
a list with two or four elements: \describe{
\item{T}{A vector of \eqn{m} test statistics}
\item{T0}{A \eqn{m \times B} matrix of permutation test statistics}
\item{p}{A vector of \eqn{m} \eqn{p}-values (only if \code{p.value} is @TRUE)}
\item{p0}{A \eqn{m \times B} matrix of permutation \eqn{p}-values (ony if \code{p.value} is @TRUE)}
}
}
\description{
Perform permutation testing for equal means in two groups
}
\details{
The test statistics is Welch's two-sample test for unequal variances
}
\examples{

p <- 513
n <- 38
mat <- matrix(rnorm(p*n), ncol=n)
cls <- rep(c(0, 1), times=c(27, n-27))
resPerm <- testByTwoSamplePermutation(X=mat, cls=cls, B=1000, seed=123)

alpha <- 0.05
res <- jointFWERControl(resPerm$T0, refFamily="Simes", alpha=alpha, stat=-resPerm$T)

}
\references{
Ge, Y., Dudoit, S. and Speed, T.P., 2003. Resampling-based
  multiple testing for microarray data analysis. _Test_, 12(1), pp.1-77.
}
