% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdFamily.R
\name{thresholdFamily}
\alias{thresholdFamily}
\title{thresholdFamily}
\usage{
thresholdFamily(alpha, Q)
}
\arguments{
\item{alpha}{A numeric value corresponding to \eqn{lambda(\alpha)} in BNR.}

\item{Q}{A \eqn{m} x \eqn{B} matrix of \code{B} Monte-Carlo samples of \code{m} test.
statistics under the null hypothesis, sorted by rows and then by columns.}
}
\description{
thresholdFamily
}
\author{
Gilles Blanchard, Pierre Neuvial and Etienne Roquain
}

