% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP.R
\name{TP}
\alias{TP}
\title{Upper bound for the number of true discoveries in a selection}
\usage{
TP(stat, thr)
}
\arguments{
\item{stat}{A vector of test statistics of the selected items}

\item{thr}{A vector of non-increasing \eqn{K} JER-controlling thresholds}
}
\value{
An upper bound on the number of true discoveries in the selection
}
\description{
Upper bound for the number of true discoveries in a selection
}
