% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upperBoundFP.R
\name{upperBoundFP}
\alias{upperBoundFP}
\title{Upper bound for the number of false discoveries}
\usage{
upperBoundFP(stat, thr, flavor = c("BNR2016", "Mein2006", "BNR2014"))
}
\arguments{
\item{stat}{ordered test statistics}

\item{thr}{JFWER thresholds}

\item{flavor}{The algorithm to compute the bound 'BNR2014' and
'BNR2016' give identical resutls and should be slightly better
than 'Mein2006' (example situation?). Flavor 'BNR2016' has a
linear time complexity, hence it is much faster than 'Mein2006'
and much much faster than 'BNR2014'.}
}
\value{
An upper bound on the number of false discoveries
}
\description{
Upper bound for the number of false discoveries
}
\author{
Gilles Blanchard, Pierre Neuvial and Etienne Roquain
}
