% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{V.star.all.leafs}
\alias{V.star.all.leafs}
\title{Post hoc bound on the number of false positives}
\usage{
V.star.all.leafs(S, C, ZL, leaf_list)
}
\arguments{
\item{S}{A subset of hypotheses}

\item{C}{Complete tree structure}

\item{ZL}{Complete zeta tree}

\item{leaf_list}{List of leaves}
}
\value{
An integer value, upper bound on the number false positives in S
}
\description{
Post hoc bound on the number of false positives
}
