% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Volcano plot}
\usage{
volcanoPlot(x, ...)
}
\arguments{
\item{x}{An object of class \code{SansSouci}, or alternatively a numeric vector
of fold changes when calling \code{volcanoPlot.numeric} directly (x axis of
the volcano plot). See \code{\link[=volcanoPlot.numeric]{volcanoPlot.numeric()}} or \code{\link[=volcanoPlot.SansSouci]{volcanoPlot.SansSouci()}}
for specifics}

\item{...}{Other arguments passed to methods}
}
\description{
Volcano plot for differential expression studies
}
