% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxStat.R
\name{wilcoxStat}
\alias{wilcoxStat}
\title{wilcoxStat}
\usage{
wilcoxStat(X, y, B = 1000)
}
\arguments{
\item{X}{An \eqn{m x n} covariate matrix}

\item{y}{A vector of \eqn{n} phenotypes in \eqn{{0,1}}}

\item{B}{Number of resamplings}
}
\value{
A list with elements \describe{
  \item{stat}{A vector of \code{m} Wilcoxon sum rank test statistics of association between \code{X} and \code{y}.}
  \item{stat0Mat}{An \code{m} x \code{B} matrix of \code{B} realizations of a \code{m}-dimensional vector of test statistics under the null hypothesis of no association between \code{X} and \code{y}.}}
}
\description{
Calculation of Wilcoxon sum rank test statistic
}
\examples{

m <- 123
rho <- 0.2
n <- 100
pi0 <- 0.5
B <- 1e2

## two-sample data
sim <- gaussianSamples(m, rho, n, pi0, SNR = 2, prob = 0.5)
cls <- rep(0, n)
cls[which(colnames(sim$X)=="1")] <- 1
stats <- wilcoxStat(sim$X, cls, 1)
stats2 <- sapply(1:m, FUN=function(ii) wilcox.test(sim$X[ii, cls==0], sim$X[ii, cls==1])$statistic)
}
\author{
Gilles Blanchard, Pierre Neuvial and Etienne Roquain
}
