% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{zeta.HB}
\alias{zeta.HB}
\title{Estimate the proportion of true nulls from a Holm-Bonferroni test}
\usage{
zeta.HB(pval, lambda)
}
\arguments{
\item{pval}{A vector of eqn{p}-values}

\item{lambda}{A numeric value in \eqn{[0,1]}, the target level of the test}
}
\description{
Estimate the proportion of true nulls from a Holm-Bonferroni test
}
