% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-functions.R
\name{zeta}
\alias{zeta}
\alias{zeta.HB}
\alias{zeta.trivial}
\alias{zeta.DKWM}
\title{Estimate the number of true null hypotheses among a set of p-values}
\usage{
zeta.HB(pval, lambda)

zeta.trivial(pval, lambda)

zeta.DKWM(pval, lambda)
}
\arguments{
\item{pval}{A vector of \eqn{p}-values}

\item{lambda}{A numeric value in \eqn{[0,1]}, the target error level of the test}
}
\value{
The number of true nulls is over-estimated as follows:
\describe{
\item{\code{zeta.DKWM}}{Inversion of the Dvoretzky-Kiefer-Wolfowitz-Massart inequality (related to the Storey estimator of the proportion of true nulls) with parameter \code{lambda}}
\item{\code{zeta.HB}}{Number of conserved hypotheses of the Holm-Bonferroni procedure with parameter \code{lambda}}
\item{\code{zeta.trivial}}{The size of the p-value set which is the trivial upper bound (\eqn{lambda} is not used)}
}
}
\description{
An upper bound on the number of true null hypotheses in the region associated to
the \eqn{p}-values \code{pval}
is computed with confidence \code{1 - lambda}.
The functions described here can be used as the \code{method} argument
of \code{\link[=zetas.tree]{zetas.tree()}}.
}
\examples{
x <- rnorm(100, mean = c(rep(c(0, 2), each = 50)))
pval <- 1 - pnorm(x)
lambda <- 0.05
zeta.trivial(pval, lambda)

zeta.HB(pval, lambda)

zeta.DKWM(pval, lambda)
}
\references{
Durand, G., Blanchard, G., Neuvial, P., & Roquain, E. (2020). Post hoc false positive control for structured hypotheses. Scandinavian Journal of Statistics, 47(4), 1114-1148.

Dvoretzky, A., Kiefer, J., and Wolfowitz, J. (1956). Asymptotic minimax character of the sample distribution function and of the classical multinomial estimator. The Annals of Mathematical Statistics, pages 642-669.

Holm, S. A simple sequentially rejective multiple test procedure. Scandinavian Journal of Statistics 6 (1979), pp. 65-70.

Massart, P. (1990). The tight constant in the Dvoretzky-Kiefer-Wolfowitz inequality. The Annals of Probability, pages 1269-1283.

Storey, J. D. (2002). A direct approach to false discovery rates. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 64(3):479-498.
}
