% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{zetas.tree.refined}
\alias{zetas.tree.refined}
\title{Estimate of the proportion of true nulls in each node of a tree}
\usage{
zetas.tree.refined(C, leaf_list, method, pvalues, alpha)
}
\description{
Estimate of the proportion of true nulls in each node of a tree
}
\details{
In this refined version, one tries to estimate the number of regions containing only signal
}
