% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SansSouci-class.R, R/SansSouciStruct-class.R
\name{all-generics}
\alias{all-generics}
\alias{nHyp}
\alias{nObs}
\alias{label}
\alias{pValues}
\alias{foldChanges}
\alias{thresholds}
\alias{nLeaves}
\title{Generic functions for S3 class SansSouci}
\usage{
nHyp(object)

nObs(object)

label(object)

pValues(object)

foldChanges(object)

thresholds(object)

nLeaves(object)
}
\arguments{
\item{object}{An object. See individual methods for specifics}
}
\description{
Generic functions for S3 class SansSouci
}
\section{Functions}{
\itemize{
\item \code{nHyp()}: ' Get the number of hypotheses

\item \code{nObs()}: Get the number of observations

\item \code{label()}: Get the label of hypotheses tested

\item \code{pValues()}: Get p-values

\item \code{foldChanges()}: Get fold changes

\item \code{thresholds()}: Get thresholds

\item \code{nLeaves()}: Get the number of leaves

}}
