% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boots_param}
\alias{boots_param}
\title{Title}
\usage{
boots_param(
  A_est,
  Pi_est,
  x_from,
  prob1,
  h,
  Sel_from,
  al,
  seuil,
  min_size,
  n,
  max_pi0,
  m0_init,
  sd0_init,
  df_init,
  norm_init,
  type_init
)
}
\arguments{
\item{A_est}{the  transition matrix}

\item{Pi_est}{initial probabilities (if null the stationnary distributio  is used)}

\item{x_from}{the observed value}

\item{prob1}{the probability of being in state one ( obtain using forward backward algorithm for instance)}

\item{h}{the size of the window in the kernel algorithm}

\item{Sel_from}{the selection tibble for theobserved value (output of selection_tibble )}

\item{al}{the given risk}

\item{seuil}{threshold for the pvalues}

\item{min_size}{the minimum size of concomitant selected position in S ( by default ony used for viterbi_min_size)}

\item{n}{the number of individuals}

\item{max_pi0}{a given maximum values of pi_0}

\item{m0_init}{if the "guess" distribution of x under the nulll state is the normal this is the guess value of the mean}

\item{sd0_init}{if the "guess" distribution of x under the nulll state is the normal this is the guess value of the standard}

\item{df_init}{if the distribution given to the EM is student, this is the degreee of freedom}

\item{norm_init}{logical, is the distribution under the null given to the EM algo is gaussian or not ? (if not consider as student)}
}
\value{

}
\description{
Title
}
