% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SansSouci-class.R, R/calibrateJER.R
\name{fit.SansSouci}
\alias{fit.SansSouci}
\alias{calibrateJER}
\title{Calibration of joint Family-Wise Error Rate thresholds}
\usage{
\method{fit}{SansSouci}(
  object,
  alpha,
  B = ceiling(10/alpha),
  rowTestFUN = NULL,
  alternative = c("two.sided", "less", "greater"),
  family = c("Simes", "Beta", "Oracle"),
  max_steps_down = 10L,
  K = nHyp(object),
  flavor = c("v1", "v0"),
  force = FALSE,
  verbose = FALSE,
  ...
)

calibrateJER(
  X,
  categ,
  B,
  alpha,
  alternative = c("two.sided", "less", "greater"),
  rowTestFUN = rowWelchTests,
  refFamily = c("Simes", "kFWER", "Beta"),
  maxStepsDown = 10L,
  K = nrow(X),
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of class \code{SansSouci}}

\item{alpha}{Target risk (JER) level}

\item{B}{A numeric value, the number of permutations to be performed}

\item{rowTestFUN}{A (vectorized) test function. Defaults to
\code{\link{rowWelchTests}}}

\item{alternative}{A character string specifying the alternative hypothesis.
Must be one of "two.sided" (default), "greater" or "less".}

\item{family}{A character value which can be \describe{
\item{Simes}{The classical family of thresholds introduced by Simes (1986):
\eqn{\alpha*k/m}. This family yields JER control if the test
statistics are positively dependent (PRDS) under H0.}

\item{Beta}{A family of thresholds that achieves "balanced" JER control under independence}

\item{Oracle}{A family such that the associated bounds correspond to the true numbers/proportions of true/false positives. "truth" must be available in object$input$truth.}
}}

\item{K}{For JER control over \code{1:K}, ie joint control of all
\eqn{k}-FWER, \eqn{k \le K}. Defaults to \eqn{m}.}

\item{flavor}{A character value (for internal use)}

\item{force}{A boolean value: should the permutation p-values and pivotal statistics be re-calculated ? Defaults to \code{FALSE}}

\item{verbose}{A boolean value: should extra info be printed?}

\item{...}{Not used}

\item{X}{A matrix of \eqn{m} variables (hypotheses) by \eqn{n} observations}

\item{categ}{An optional numeric vector of \eqn{n} values in \eqn{0, 1}
specifying the column indices of the first and second samples. If not
provided, a one-sample test is performed.}

\item{refFamily}{A character value which can be \describe{

\item{Simes}{The classical family of thresholds introduced by Simes (1986):
\eqn{\alpha*k/m}. This family yields joint FWER control if the test
statistics are positively dependent (PRDS) under H0.}

\item{Beta}{A family of thresholds that achieves "balanced" joint error
rate (JER) control under independence}

\item{kFWER}{A family \eqn{(t_k)} calibrated so that for each k,
\eqn{(t_k)} controls the (marginal) k-FWER.}}}

\item{maxStepsDown}{Maximum number of steps down to be performed.}
}
\value{
A 'fitted' object of class 'SansSouci'. It is a list of three elements
\itemize{
\item input: see \link{SansSouci}
\item param: the input parameters, given as a list
\item output: the outputs of the calibration, see \link{calibrateJER}
}

A list with elements: \describe{

\item{p.values}{A numeric vector of \code{m} p-values}
\item{pivStat}{A numeric vector of length \code{B}, the values of the pivotal
statistic whose quantile of order \eqn{alpha} is \eqn{lambda}.}
\item{thr}{A numeric vector of length \code{K}, such that the estimated
probability that there exists an index \eqn{k} between 1 and \eqn{K} such
that the \eqn{k}-th maximum of the test statistics of is greater than
\eqn{thr[k]}, is less than \eqn{\alpha}}

\item{lambda}{A numeric value, the result of the calibration}

\item{FP}{A numeric vector of length \code{m}, a 1-alpha confidence bound on the number of false positives}
}
}
\description{
Calibration of of JER thresholds using one or two-sample tests
}
\details{
See \code{\link{testByRandomization}} for a description of the tests performed for calibration.
}
\section{Methods (by generic)}{
\itemize{
\item \code{fit}: Fit SansSouci object
}}

\examples{

m <- 543
pi0 <- 0.8
sim <- gaussianSamples(m = m, rho = 0.4, n = 100,
                       pi0 = pi0, SNR = 3, prob = 0.5)
X <- sim$X
categ <- sim$categ
alpha <- 0.1
cal <- calibrateJER(X, categ, B = 1e2, alpha = alpha, refFamily="Simes")
cal$lambda # > alpha (whp) if rho > 0

# Application 1: confidence bounds
#   ie upper confidence bound for the number of false positives 
#   among the k most significant items for all k
cb <- cal$conf_bound
plotConfCurve(cb, xmax = 200)
  
## Compare to Simes (without calibration) and "Oracle" (ie truth from the simulation settings)
cb_Simes <- confCurveFromFam(cal$p.values, refFamily = "Simes", param = alpha)
cb_Oracle <- confCurveFromFam(cal$p.values, refFamily = "Oracle", param = (sim$H == 1))
all_cb <- list("Simes + calibration" = cb, 
                "Simes"= cb_Simes, 
                "Oracle" = cb_Oracle)
plotConfCurve(all_cb, xmax = 200)

# Application 2a: bound on the number of false positives in one or 
#    more user-defined selections

spval <- sort(cal$p.values)
sel <- spval[c(1:10)]
maxFP(sel, cal$thr)

sel <- spval[c(1:10, 35:40)]
maxFP(sel, cal$thr)

sel <- spval[c(30:50)]
maxFP(sel, cal$thr)

# Application 2b: bound on pi0, the proportion of false positives in H

maxFP(spval, cal$thr)/m
pi0

}
\references{
Blanchard, G., Neuvial, P., & Roquain, E. (2020). Post hoc confidence bounds on false positives using reference families. Annals of Statistics, 48(3), 1281-1303.
}
\author{
Gilles Blanchard, Pierre Neuvial and Etienne Roquain
}
