% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_env_asupr.R
\name{confidenceEnvelope}
\alias{confidenceEnvelope}
\title{Title}
\usage{
confidenceEnvelope(
  p.values,
  refFamily,
  param,
  K = length(p.values),
  what = c("TP", "FDP")
)
}
\arguments{
\item{p.values}{}

\item{refFamily}{}

\item{param}{}

\item{K}{}

\item{what}{}
}
\value{

}
\description{
Title
}
