% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{dichotomy}
\alias{dichotomy}
\title{For a given risk level, calibrate the method on learned templates by
dichotomy. This is equivalent to calibrating using pivotal stats but does
not require the availability of a closed form inverse template.}
\usage{
dichotomy(alpha, learned_templates, pval0, k_max, min_dist = 3)
}
\arguments{
\item{alpha}{A float, confidence level in \link{0, 1}}

\item{learned_templates}{A matrix with B rows and p columns. Learned templates for B' permutations and p features}

\item{pval0}{A matrix with B rows and p columns of permuted p-values.}

\item{k_max}{An integer value, the template size.}

\item{min_dist}{An integer, minimum distance to stop iterating dichotomy. Defaults to \code{m}}
}
\value{
A integer, index of template chosen by calibration
}
\description{
For a given risk level, calibrate the method on learned templates by
dichotomy. This is equivalent to calibrating using pivotal stats but does
not require the availability of a closed form inverse template.
}
\examples{

m <- 50
n <- 45
X <- matrix(rnorm(m*n), ncol = n, nrow = m)
categ <- rbinom(n, 1, 0.4)
B <- 10
null_groups <- replicate(B, sample(categ))
p0 <- rowWelchTests(X, null_groups)$p.value
pivStat <- get_pivotal_stat(p0, m)
quantile(pivStat, 0.2)


}
\references{
Blanchard, G., Neuvial, P., & Roquain, E. (2020). Post hoc
confidence bounds on false positives using reference families. \emph{Annals of
Statistics}, 48(3), 1281-1303.
}
