% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{dyadic}
\alias{dyadic}
\alias{dyadic.from.window.size}
\alias{dyadic.from.height}
\alias{dyadic.from.max.height}
\title{Create a complete dyadic tree structure}
\usage{
dyadic.from.window.size(m, s, method)

dyadic.from.height(m, H, method)

dyadic.from.max.height(m, method)
}
\arguments{
\item{m}{An integer value, the number of elements in the structure}

\item{s}{An integer value, the number of elements in each leaf}

\item{method}{A numeric value. If \code{method == 1}, start from the leaves
and group nodes of a same height 2 by 2 as long as possible. If
\code{method==2}, start from the root and divide nodes in 2 nodes of equal
size as long as possible}

\item{H}{An integer value, the desired maximal height of the tree}
}
\value{
A list with two elements:\describe{
\item{leaf_list}{A list of leaves}
\item{C}{A list of lists containing the dyadic structure}
}
}
\description{
Create a complete dyadic tree structure
}
\details{
\describe{
\item{\code{dyadic.from.window.size}}{Dyadic tree structure from window size: the number of elements in each leaf is set to \code{s}}
\item{\code{dyadic.from.height}}{Dyadic tree structure from height: the total height of tree is set to \code{H}}
\item{\code{dyadic.from.max.size}}{Dyadic tree structure from maximum achievable height: the total height of tree is set to the maximum possible height, ie \code{floor(2 + log2(m - 1))}}
}
}
\examples{
m <- 6
dd <- dyadic.from.window.size(m, s = 2, method = 2)
str(dd)

dd <- dyadic.from.height(m, H = 3, method = 2)
str(dd)

dd <- dyadic.from.max.height(m, method=2)
str(dd)

leaf_list <- dd$leaf_list
}
