% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empiricalCoverage.R
\name{empiricalCoverage}
\alias{empiricalCoverage}
\title{Empirical coverage of a threshold family}
\usage{
empiricalCoverage(thr, mat)
}
\arguments{
\item{thr}{A numeric vector of length \code{m}, a threshold family. Should  be sorted in decreasing order.}

\item{mat}{A \eqn{c} x \eqn{B} matrix of \code{B} Monte-Carlo
samples of \code{c} test statistics under the null hypothesis.}
}
\value{
The empirical coverage of \code{thr} with respect to \code{mat}.
}
\description{
Empirical coverage of a threshold family
}
\examples{
alpha <- 0.05

# empirical coverage of Simes thresholds for independent Gaussian nulls
m <- 234
B <- 1234
X0 <- replicate(B, rnorm(m))
thr <- SimesThresholdFamily(m)(alpha)
sansSouci:::empiricalCoverage(thr, X0)

}
