% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SansSouciStruct-class.R
\name{fit.SansSouciStruct}
\alias{fit.SansSouciStruct}
\title{Fit SansSouciStruct object}
\usage{
\method{fit}{SansSouciStruct}(
  object,
  alpha,
  p.values,
  family = c("DKWM", "HB", "trivial", "Simes", "Oracle"),
  flavor = c("tree", "partition"),
  ...
)
}
\description{
Fit SansSouciStruct object
}
\examples{
s <- 100
q <- 7
m <- s*2^q
obj <- SansSouciDyadic(m, s, flavor = "window.size", direction = "top-down")

mu <- gen.mu.leaves(m = m, K1 = 8, d = 0.9, grouped = TRUE, 
  setting = "const", barmu = 3, leaf_list = obj$input$leaves)
pvalues <- gen.p.values(m = m, mu = mu, rho = 0)

alpha <- 0.05
res <- fit(obj, alpha, pvalues, "DKWM")
bound(res)
}
