% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{get_pivotal_stat}
\alias{get_pivotal_stat}
\title{Get a vector of pivotal statistics associated
to permutation p-values and to a reference family}
\usage{
get_pivotal_stat(p0, t_inv = t_inv_linear, m = nrow(p0), K = m)
}
\arguments{
\item{p0}{A \verb{m x B} matrix. The j-th ow corresponds to a permutation
of the input \code{categ}: for each hypothesis i, \code{p0[i,j]} is the
p-value of the test of the i-th null hypothesis on the permuted categories}

\item{t_inv}{An inverse threshold function (same I/O as \link{t_inv_linear})}

\item{m}{The total numer of tested hypotheses. Defaults to \code{p0}}

\item{K}{An integer value in \verb{[1,m]}, the number of elements in the reference family. Defaults to \code{m}}

\item{alpha}{A numeric value in \verb{[0,1]}, the target risk}

\item{family}{A character value, the name of a threshold family. Should be
one of "Linear", "Beta" and "Simes". "Linear" and "Simes" families are
identical.}
}
\value{
A vector of length \code{B} pivotal statitics, whose j-th entry
corresponds to \code{psi(g_j.X)} with notation of the AoS 2020 paper cited
below (section 4.5)
}
\description{
Get a vector of pivotal statistics associated
to permutation p-values and to a reference family
}
\examples{

m <- 50
n <- 45
X <- matrix(rnorm(m*n), ncol = n, nrow = m)
categ <- rbinom(n, 1, 0.4)
B <- 10
null_groups <- replicate(B, sample(categ))
p0 <- rowWelchTests(X, null_groups)$p.value
pivStat <- get_pivotal_stat(p0)
quantile(pivStat, 0.2)


}
\references{
Blanchard, G., Neuvial, P., & Roquain, E. (2020). Post hoc
confidence bounds on false positives using reference families. \emph{Annals of
Statistics}, 48(3), 1281-1303.
}
