% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointFWERControl.R
\name{jointFWERControl}
\alias{jointFWERControl}
\title{Calibration of joint Family-Wise error rate thresholds}
\usage{
jointFWERControl(mat, refFamily = c("Simes", "kFWER"), alpha, stat = NULL,
  maxStepsDown = 100, kMax = nrow(mat), Rcpp = TRUE, verbose = TRUE)
}
\arguments{
\item{mat}{A \eqn{m} x \eqn{B} matrix of Monte-Carlo samples of
test statistics under the null hypothesis. \describe{
\item{m}{is the number of tested hypotheses} \item{B}{is the
number of Monte-Carlo samples}}}

\item{refFamily}{A character value which can be \describe{
\item{Simes}{The classical family of thresholds introduced by
Simes (1986): \eqn{\alpha*k/m}. This family yields joint FWER
control if the test statistics are positively dependent (PRDS)
under H0.}
\item{kFWER}{A family \eqn{(t_k)} calibrated so that for each k,
\eqn{(t_k)} controls the (marginal) k-FWER.}}}

\item{alpha}{Target joint FWER level.}

\item{stat}{A vector of \eqn{m} test statistics. Not used for
single step control, and mandatory for step-down JFWER control. If
not provided, single step control is performed.}

\item{maxStepsDown}{Maximum number of steps down to be performed.
\code{maxSteps=1} corresponds to single step JFWER control.}

\item{kMax}{For simultaneous control of (\eqn{k}-FWER for all
\eqn{k \le k[max]}).}

\item{Rcpp}{If \code{TRUE}, some costly operations (sorting) are
performed in C++.}

\item{verbose}{A boolean value: should extra info be printed?}
}
\value{
A list with elements: \describe{
    \item{thr}{A numeric vector of length \code{m}, such that the
    estimated probability that there exists an index \eqn{k} between 1
    and m such that the k-th maximum of the test statistics of is
    greater than \eqn{thr[k]}, is less than \eqn{\alpha}.}
    \item{pivStat}{A numeric vector of length \code{m}, the values of the pivotal
    statistic whose quantile of order \eqn{alpha} is \eqn{lambda}.}
    \item{lambda}{JFWER threshold.}
    \item{Vbar}{An upper bound on the number of false discoveries,
    as calculated by \code{curveMaxFP(stat, thr)}.}
    \item{steps}{a list with elements named 'thr', 'pivStat' and
    'lambda' giving the sequence of corresponding vectors/values
    along the steps down.}}
}
\description{
Calibration of a family of thresholds that provide joint FWER control
}
\examples{

set.seed(0xBEEF)
m <- 123
rho <- 0.2
n <- 100
pi0 <- 0.5
B <- 1e3

sim <- gaussianTestStatistics(m, B, pi0 = pi0, SNR = 3)
X0 <- sim$X0
x <- sim$x

## Test statistics
pch <- 20
plot(x, col=1+sim$H, main="Test statistics", pch=pch)
legend("topleft", c("H0", "H1"), pch=pch, col=1:2)

alpha <- 0.1
res <- jointFWERControl(X0, refFamily="kFWER", alpha=alpha, stat=x)
Vbar <- res$Vbar
ttl <- paste("Upper bound on the number of false positives", "among first k hypotheses", sep="\\n")
plot(Vbar, main=ttl, xlab="k", ylab=expression(bar(V(k))))

}
\author{
Gilles Blanchard, Pierre Neuvial and Etienne Roquain
}
