% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapCalibration.R
\name{lm_test}
\alias{lm_test}
\title{Mass-univariate inference for contrasts in a linear model}
\usage{
lm_test(Y, X, C, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{Y}{A data matrix of size $n$ observations (in row) and $D$ features in
columns}

\item{X}{A design matrix of size $n$ observations (in row) and $p$ variables
(in columns)}

\item{C}{A contrast matrix of size $L$ tested contrasts (in row) and $p$
columns corresponding to the parameters to be tested}

\item{alternative}{A character string specifying the alternative hypothesis.
Must be one of "two.sided" (default), "greater" or "less".}
}
\value{
A list with elements:
\describe{
\item{epsilon_est}{A \eqn{n \times D} matrix of residuals}
\item{stat_test}{A \eqn{L \times D} matrix of test statistics}
\item{p.value}{A \eqn{L \times D} matrix of p-values}
\item{beta_est}{A \eqn{n \times D} matrix of parameter estimates}
}
}
\description{
Compute the marginal t-statistics for a set of contrasts and their
(two-sided) p-value.
}
\details{
Based on a python implementation available in the \code{pyperm}
package: \url{https://github.com/sjdavenport/pyperm/}
}
\examples{
N <- 100
p <- 2
D <- 2
X <- matrix(0, nrow = N, ncol = p)
X[, 1] <- 1
X[, -1] <- runif(N*(p-1), min = 0, max = 3)
beta <- matrix(0, nrow = p, ncol = D)
epsilons <- matrix(rnorm(N*D), nrow = N, ncol = D)
Y <- X \%*\% beta + epsilons
C <- diag(p)
resLM <- lm_test(Y = Y, X = X, C = C)
}
\references{
Davenport, S., Thirion, B., & Neuvial, P. (2025). FDP control in
mass-univariate linear models using the residual bootstrap. Electronic
Journal of Statistics, 19(1), 1313-1336.
}
