% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc-bounds.R
\name{maxFP}
\alias{maxFP}
\title{(Slow) Upper bound for the number of false discoveries in a selection}
\usage{
maxFP(p.values, thr)
}
\arguments{
\item{p.values}{A vector of p-values for the selected items}

\item{thr}{A vector of non-decreasing k-FWER-controlling thresholds}
}
\value{
A post hoc upper bound on the number of false discoveries in the selection
}
\description{
(Slow) Upper bound for the number of false discoveries in a selection
}
\examples{

m <- 123
sim <- gaussianSamples(m = m, rho = 0.2, n = 100, 
                       pi0 = 0.8, SNR = 3, prob = 0.5)
X <- sim$X
groups <- sim$categ
p <- rowWelchTests(X, groups)$p.value

null_groups <- replicate(100, sample(groups))
p0 <- rowWelchTests(X, null_groups)$p.value
calib <- calibrate(p0, alpha = 0.1)
thr <- calib$thr

M0 <- maxFP(p, thr)
M0/m

maxFP(head(p), thr)
maxFP(tail(p), thr)
maxFP(c(head(p), tail(p)), thr)

}
\references{
Blanchard, G., Neuvial, P., & Roquain, E. (2020). Post hoc confidence bounds on false positives using reference families. Annals of Statistics, 48(3), 1281-1303.
}
