% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc_bounds.R
\name{maxFP}
\alias{maxFP}
\title{Upper bound for the number of false discoveries in a selection}
\usage{
maxFP(p.values, thr)
}
\arguments{
\item{p.values}{A vector of p-values for the selected items}

\item{thr}{A vector of non-decreasing k-FWER-controlling thresholds}
}
\value{
A post hoc upper bound on the number of false discoveries in the selection
}
\description{
Upper bound for the number of false discoveries in a selection
}
\examples{

m <- 123
sim <- gaussianSamples(m = m, rho = 0.2, n = 100, 
                       pi0 = 0.8, SNR = 2.5, prob = 0.5)
X <- sim$X
cal <- calibrateJER(X, sim$categ, B = 1e3, alpha = 0.2, refFamily="Simes", )
thr <- sort(cal$thr)
pval <- sort(cal$p.values)

M0 <- maxFP(pval, cal$thr) ## upper bound on m0...
M0/m

maxFP(head(pval), thr)
maxFP(tail(pval), thr)
maxFP(c(head(pval), tail(pval)), thr)

}
\references{
Blanchard, G., Neuvial, P., & Roquain, E. (2020). Post hoc confidence bounds on false positives using reference families. Annals of Statistics, 48(3), 1281-1303.
}
