% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_batch_rowTestFUN.R
\name{mini_batch_rowTestFUN}
\alias{mini_batch_rowTestFUN}
\title{Mini batch version of a rowTestFUN function.}
\usage{
mini_batch_rowTestFUN(
  rowTestFUN,
  Y,
  categ,
  alternative = c("two.sided", "less", "greater"),
  max_batch_size = 1e+06
)
}
\arguments{
\item{rowTestFUN}{a function taking as inputs Y, categ and Alternative and
giving as a result a matrix of $p$-values. Example \code{\link{rowWelchTests}} or
\code{\link{rowWilcoxonTests}}.}

\item{Y}{a $m \times n$ numeric matrix whose rows correspond to variables
and columns to observations}

\item{categ}{}

\item{alternative}{}

\item{max_batch_size}{}
}
\value{
the p-values for the test
}
\description{
Mini batch version of a rowTestFUN function.
}
\examples{
p <- 53 
n <- 35
mat <- matrix(rnorm(p * n), ncol = n)
cls <- rep(c(0, 1), times = c(10, n - 10))
mini_batch_rowTestFUN(rowTestFUN = rowWelchTests, Y = mat, 
                      categ = cls, alternative = alt, 
                      max_batch_size = 1e6)
}
