% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc_bounds.R
\name{minTP}
\alias{minTP}
\title{Lower bound for the number of true discoveries in a selection}
\usage{
minTP(p.values, thr)
}
\arguments{
\item{p.values}{A vector of p-values for the selected items}

\item{thr}{A vector of non-decreasing k-FWER-controlling thresholds}
}
\value{
A Lower bound on the number of true discoveries in the selection
}
\description{
Lower bound for the number of true discoveries in a selection
}
