% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SansSouci-class.R
\name{plot.SansSouci}
\alias{plot.SansSouci}
\title{Plot confidence bound on the true/false positives among most significant items}
\usage{
\method{plot}{SansSouci}(x, y, xmax = nHyp(x), ...)
}
\arguments{
\item{x}{An object of class 'SansSouci'}

\item{y}{Not used}

\item{xmax}{Right limit of the plot}

\item{...}{Further arguments to be passed to \code{bound}}
}
\description{
Plot confidence bound on the true/false positives among most significant items
}
\examples{
# Generate Gaussian data and perform multiple tests
obj <- SansSouciSim(m = 502, rho = 0.5, n = 100, pi0 = 0.8, SNR = 3, prob = 0.5)
res <- fit(obj, B = 100, alpha = 0.1)

# confidence curve
plot(res)

# confidence curve for a subset of hypotheses
S <- which(pValues(res) < 0.1 & foldChanges(res) > 0.3)
plot(res, S = S)
}
