% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc-bounds.R
\name{plotConfCurve}
\alias{plotConfCurve}
\title{Plot confidence bound}
\usage{
plotConfCurve(conf_bound, xmax, cols = NULL)
}
\arguments{
\item{conf_bound}{A data.frame or a list of data.frames as output by \link{predict}}

\item{xmax}{Right limit of the plot}

\item{cols}{A vector of colors of the same length as \code{conf_bound}}
}
\description{
Plot confidence bound
}
\references{
Blanchard, G., Neuvial, P., & Roquain, E. (2020). Post hoc confidence bounds on false positives using reference families. Annals of Statistics, 48(3), 1281-1303.
}
