% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthocBySimes.R
\name{posthocBySimes}
\alias{posthocBySimes}
\title{posthocBySimes}
\usage{
posthocBySimes(p, select, alpha = 0.05, verbose = FALSE)
}
\arguments{
\item{p}{A numeric vector of \code{m} p-values for all tested hypotheses.}

\item{select}{A vector of indices in \eqn{[1, \dots m]} of the hypotheses to be selected.}

\item{alpha}{A numeric value, the significance level of the test procedure.}

\item{verbose}{If \code{TRUE}, prints verbose result to the
screen. Defaults to \code{FALSE}.}
}
\value{
A integer value, Simes's lower bound on the number of
 correct rejections within the selected hypotheses
}
\description{
Lower bound on the number of correct rejections using Simes' threshold
family
}
\details{
If (R_k)_k provides jFWER control at level \eqn{\alpha} then with
probability greater than \eqn{1-\alpha}, \eqn{|H_0 cap R| \leq \min_k {|R
\cap (R_k)^c|+k-1}} A bit better: \eqn{|H_0 cap R| \leq (\min_{k<= |R|} {|R
\cap R_k^c|+k-1}) \wedge |R|}
}
\examples{
m <- 1e3
m1 <- 200
p <- 1-pnorm(c(rnorm(m1, mean=4), rnorm(m-m1, mean=0)))
R <- sample(m, 10)
if (require("cherry")) {
  hom <- hommelFast(p)
  pickSimes(hom, R, silent=TRUE)
}
posthocBySimes(p, R)
}
\author{
Gilles Blanchard, Pierre Neuvial and Etienne Roquain
}

