% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/posthocBySimes.R
\name{posthocBySimes0Rcpp}
\alias{posthocBySimes0Rcpp}
\alias{posthocBySimes}
\title{post hoc bound obtained from Simes' inequality}
\usage{
posthocBySimes0Rcpp(p, select, alpha)

posthocBySimes(p, select, alpha, Rcpp = FALSE, verbose = FALSE)
}
\arguments{
\item{p}{A numeric vector of \code{m} p-values for all tested hypotheses.}

\item{select}{A vector of indices in \eqn{[1, \dots m]} of the hypotheses to be selected.}

\item{alpha}{A numeric value, the significance level of the test procedure.}

\item{Rcpp}{A boolean value: use Rcpp version of the implementation? Defaults to FALSE.}

\item{verbose}{If \code{TRUE}, prints verbose result to the
screen. Defaults to \code{FALSE}.}
}
\value{
A integer value, Simes's lower bound on the number of
 correct rejections within the selected hypotheses
}
\description{
Lower bound on the number of correct rejections using Simes' reference
family
}
\details{
If (R_k)_k provides jFWER control at level \eqn{\alpha} then with
probability greater than \eqn{1-\alpha}, \eqn{|H_0 cap R| \leq \min_k {|R
\cap (R_k)^c|+k-1}} A bit better: \eqn{|H_0 cap R| \leq (\min_{k<= |R|} {|R
\cap R_k^c|+k-1}) \wedge |R|}
}
\section{Functions}{
\itemize{
\item \code{posthocBySimes0Rcpp}: Rcpp version

\item \code{posthocBySimes}: R version
}}

\examples{
m <- 1e3
m1 <- 200
p <- 1-pnorm(c(rnorm(m1, mean=4), rnorm(m-m1, mean=0)))
R <- union(1:10, sample(m, 10))
alpha <- 0.10
if (require("cherry")) {
  hom <- hommelFast(p)
  pickSimes(hom, R, silent=TRUE, alpha = alpha)
}
posthocBySimes(p, R, alpha=alpha)

}
\references{
Blanchard, G., Neuvial, P., & Roquain, E. (2020). Post hoc confidence bounds on false positives using reference families. Annals of Statistics, 48(3), 1281-1303.

Goeman, J. J., & Solari, A. (2011). Multiple testing for exploratory research. Statistical Science, 26(4), 584-597.
}
\author{
Gilles Blanchard, Pierre Neuvial and Etienne Roquain
}
