% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowWelchTests.R
\name{rowWelchTests}
\alias{rowWelchTests}
\title{Welch t-tests for rows of a matrix}
\usage{
rowWelchTests(mat, categ, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{mat}{A numeric matrix whose rows correspond to variables and columns to
observations}

\item{categ}{A vector of \code{ncol(mat)} categories in \eqn{'0','1'} for the
observations}

\item{alternative}{A character string specifying the alternative hypothesis.
Must be one of "two.sided" (default), "greater" or "less". As in
\code{\link{t.test}}, alternative = "greater" is the alternative that the
class '1' has a larger mean than the class '0'.}
}
\value{
A list with class "htest" containing the following components:
  \describe{ \item{statistic}{the value of the t-statistics}
  \item{parameter}{the degrees of freedom for the t-statistics}
  \item{p.value}{the p-values for the tests}} \item{meanDiff}{the mean
  difference}
}
\description{
Welch t-tests for each row of a matrix, intended to be speed efficient
}
\details{
Other functions to perform t-tests on each row of a matrix include:
\code{\link[genefilter]{fastT}}: this function only calculates the test
statistics and not the corresponding degrees of freedom or p-value;
\code{\link[genefilter]{rowttests}}: this function calculates the test
statistics, degrees of freedom and associated p-value, but only in the case
of equal group variances (standard t-test and not Welch t-test)
}
\examples{

p <- 1e3
n <- 38
mat <- matrix(rnorm(p*n), ncol=n)
cls <- rep(c(0, 1), times=c(27, n-27))
fwt <- rowWelchTests(mat, categ=cls, alternative = "greater")
str(fwt)

# compare with ordinary t.test:
pwt <- apply(mat, 1, FUN=function(x) {
   t.test(x[cls==1], x[cls==0], alternative = "greater")$p.value
})
all(abs(fwt$p.value-pwt) < 1e-10)  ## same results

}
\references{
B. L. Welch (1951), On the comparison of several mean values: an
  alternative approach. Biometrika, *38*, 330-336
}
\author{
Pierre Neuvial
}
