% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SansSouci-class.R
\name{SansSouci}
\alias{SansSouci}
\title{Create an object of class 'SansSouci'}
\usage{
SansSouci(Y, groups, truth = NULL)
}
\arguments{
\item{Y}{A matrix of \eqn{m} variables (hypotheses) by \eqn{n} observations}

\item{groups}{A numeric vector of \eqn{n} values in \eqn{0, 1}, the groups of
observations on which to perform two-sample tests}

\item{truth}{An optional numeric vector of $m$ values in ${0,1}$, the status
of each null hypothesis (0 means H0 is true, 1 means H1 is true). Typically
used in simulations.}
}
\value{
An object of class "SansSouci"
}
\description{
Create an object of class 'SansSouci'
}
\examples{
data(expr_ALL, package = "sanssouci.data")
groups <- ifelse(colnames(expr_ALL)=="NEG", 0, 1)
table(groups)
a <- SansSouci(Y = expr_ALL, groups = groups)

res <- fit(a, B = 100, alpha = 0.1)
plot(res, xmax = 500)
res_beta <- fit(res, B = 100, alpha = 0.1, family="Beta", K=10)
plot(res_beta, xmax = 500)

}
