% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SansSouciStruct-class.R
\name{SansSouciDyadic}
\alias{SansSouciDyadic}
\title{Create an object of class 'SansSouciStruct' with a dyadic structure}
\usage{
SansSouciDyadic(
  m,
  leafSize = NULL,
  maxHeight = NULL,
  flavor = c("window.size", "height", "max.height"),
  direction = c("bottom-up", "top-down"),
  ...
)
}
\arguments{
\item{m}{A numeric value, the number of hypotheses}

\item{leafSize}{A numeric value, the number of hypotheses in each leaf}

\item{maxHeight}{A numeric value, the maximum height (or depth) of the tree}

\item{flavor}{A character value defining the type of tree building function
to be used. Must be one of window.size", "height", or "max.height"}

\item{direction}{A character value, the direction used for building the tree
Must be one of "bottom-up" or "top-down"}

\item{...}{Arguments to be passed to \code{SansSouciStruct}}
}
\value{
An object of class 'SansSouciStruct'
}
\description{
Create an object of class 'SansSouciStruct' with a dyadic structure
}
\examples{
s <- 100
q <- 7
m <- s*2^q

dd <- dyadic.from.window.size(m, s, method = 2)
obj <- SansSouciDyadic(m, s, flavor = "window.size", direction = "top-down")

}
\references{
Durand, G., Blanchard, G., Neuvial, P., & Roquain, E. (2020).
Post hoc false positive control for structured hypotheses. Scandinavian
Journal of Statistics, 47(4), 1114-1148.
}
\seealso{
dyadic.from.leave.list dyadic.from.window.size dyadic.from.height
dyadic.from.max.height
}
