% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SansSouciStruct-class.R
\name{SansSouciDyadic}
\alias{SansSouciDyadic}
\title{Create an object of class 'SansSouciStruct' with a dyadic structure}
\usage{
SansSouciDyadic(
  m,
  leaf_size = NULL,
  height = NULL,
  direction = c("bottom-up", "top-down"),
  ...
)
}
\arguments{
\item{m}{A numeric value, the number of hypotheses}

\item{leaf_size}{An integer value, the number of hypotheses in each leaf.
Defaults to NULL.}

\item{height}{An integer value, the desired maximal height of the tree.
Defaults to NULL.}

\item{direction}{A character value, the direction used for building the tree.
Must be one of  "bottom-up" or "top-down"}
}
\value{
An object of class 'SansSouciStruct'
}
\description{
Create an object of class 'SansSouciStruct' with a dyadic structure
}
\details{
If both 'leaf_size' and 'height' are NULL then a binary tree with
maximal height given the total number of hypotheses is built. If both
'leaf_size' and 'height' are non NULL, then 'height' is ignored and a tree
is built with the desired leaf size.
}
\examples{
s <- 100
q <- 7
m <- s*2^q

obj <- SansSouciDyadic(m, leaf_size = s, direction = "top-down")
obj <- SansSouciDyadic(m, height = 8, direction = "top-down")

}
\references{
Durand, G., Blanchard, G., Neuvial, P., & Roquain, E. (2020).
Post hoc false positive control for structured hypotheses. Scandinavian
Journal of Statistics, 47(4), 1114-1148.
}
\seealso{
dyadic.from.leave.list dyadic.from.window.size dyadic.from.height
dyadic.from.max.height
}
