% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SansSouciStruct-class.R
\name{SansSouciStruct}
\alias{SansSouciStruct}
\title{Create an object of class 'SansSouciStruct'}
\usage{
SansSouciStruct(struct, leaves, truth = NULL)
}
\arguments{
\item{struct}{A (dyadic) structure}

\item{leaves}{A list of leaves}

\item{truth}{An optional numeric vector of $m$ values in ${0,1}$, the status of each null hypothesis (0 means H0 is true, 1 means H1 is true). Typically used in simulations.}
}
\value{
An object of class 'SansSouciStruct'
}
\description{
Create an object of class 'SansSouciStruct'
}
\examples{
s <- 100
q <- 7
m <- s*2^q

dd <- dyadic.from.window.size(m, s, method = 2)
obj <- SansSouciStruct(dd$C, dd$leaf_list)

}
\references{
Durand, G., Blanchard, G., Neuvial, P., & Roquain, E. (2020). Post hoc false positive control for structured hypotheses. Scandinavian Journal of Statistics, 47(4), 1114-1148.
}
\seealso{
SansSouciDyadic
}
