% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testByRandomization.R
\name{testByRandomization}
\alias{testByRandomization}
\title{Randomization-based testing}
\usage{
testByRandomization(X, B, cls = colnames(X), p.value = TRUE, seed = NULL)
}
\arguments{
\item{X}{a matrix of \code{m} variables by \code{n} observations}

\item{B}{A numeric value, the number of permutations to be performed}

\item{cls}{A vector of length \code{n} class labels in \code{0,1} for flavor
"perm". Defaults to colnames(X).}

\item{p.value}{A boolean value: should randomization \eqn{p}-values be
calculated and returned? Defaults to @TRUE}

\item{seed}{An integer (or NULL) value used as a seed for random number
generation. If \code{NULL}, no seed is specified}
}
\value{
a list with elements: \describe{

  \item{T}{A vector of \eqn{m} test statistics}

  \item{T0}{A \eqn{m \times B} matrix of randomized test statistics}

  \item{p}{A vector of \eqn{m} \eqn{p}-values (only if \code{p.value} is
  \code{TRUE} )}

  \item{p0}{A \eqn{m \times B} matrix of randomization \eqn{p}-values (only
  if \code{p.value} is \code{TRUE} )}

  \item{flavor}{A character value, the type of randomization performed:
  "perm" for permutation-based randomization in two-sample tests, and "flip"
  for sign-flipping-based randomization in one sample tests. See Details.}

  \item{param.p}{A vector of \eqn{m} parametric \eqn{p}-values (only for
  flavor "perm")}

  \item{param.p0}{A \eqn{m \times B} matrix of parametric \eqn{p}-values on
  permuted data (only for flavor "perm" )}

  \item{df}{A vector of \eqn{m} degrees of freedom for the observed
  statistics (only for flavor "perm")}

  \item{df0}{A \eqn{m \times B} matrix of degrees of freedom on permuted data
  (only for flavor "perm" )}}
}
\description{
Randomization-based testing using permutation or sign-flipping
}
\details{
The type of randomization is determined by the parameter \code{cls}.
  If \code{cls} does not containt two distinct values (or is \code{NULL}), a
  one-sample test is performed using randomization (flavor "flip"). If it
  contains two distinct values, a two-sample test is perfomed using
  permutations (flavor "perm").

  For permutation, we test the null hypothesis: "both groups have the same
  mean" against the one-sided alternative that the mean is larger in the
  second group. The test is Welch's two-sample test for unequal variances.
  Permuted test statistics are calculated by B permutations of the group
  labels. Corresponding observed and permuted p-values are calculated as the
  proportion of permutations (including the identity) for which the permuted
  test statistic is larger than the observed test statistic. The
  corresponding parametric p-values are returned as elements \code{param.p}
  (for the original data) and \code{param.p0} (for the permutations).

  For sign-flipping, we test the null hypothesis: "the mean is 0" against the
  two-sided alternative that the mean is larger than 0. We use the (rescaled)
  empirical mean of the observations as a test statistic. Sign-flipped test
  statistics are calculated by flipping the sign of each observation with
  probability 1/2.
}
\examples{

p <- 510
n <- 380
B <- 1e3
mat <- matrix(rnorm(p*n), ncol=n)
cls <- rep(c(0, 1), times=c(n/2, n-n/2))
resPerm <- testByRandomization(X=mat, cls=cls, B=B) ## permutation
resFlip <- testByRandomization(X=mat, B=B)          ## sign-flipping

# empirical coverage of Simes thresholds
alpha <- 0.2
thr <- SimesThresholdFamily(p)(alpha)
sansSouci:::empiricalCoverage(thr, resPerm$T0) ## Welch, not Gaussian
sansSouci:::empiricalCoverage(thr, qnorm(1-resPerm$p0))
sansSouci:::empiricalCoverage(thr, qnorm(1-resPerm$param.p0))

sansSouci:::empiricalCoverage(thr, resFlip$T0)
sansSouci:::empiricalCoverage(thr, qnorm(1-resFlip$p0))

# test statistics null distribution
hist(resPerm$p)
hist(resFlip$p)

alpha <- 0.05
resP <- jointFWERControl(resPerm$T0, refFamily="Simes", alpha=alpha, stat=resPerm$T)
resF <- jointFWERControl(resFlip$T0, refFamily="Simes", alpha=alpha, stat=resFlip$T)

}
\references{
Ge, Y., Dudoit, S. and Speed, T.P., 2003. Resampling-based
  multiple testing for microarray data analysis. _Test_, 12(1), pp.1-77.
}
