% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testByRandomization.R
\name{testByRandomization}
\alias{testByRandomization}
\title{Randomization-based testing}
\usage{
testByRandomization(
  X,
  B,
  alternative = c("two.sided", "less", "greater"),
  rowTestFUN = rowWelchTests,
  rand.p.value = FALSE,
  seed = NULL
)
}
\arguments{
\item{X}{a matrix of \code{m} variables by \code{n} observations}

\item{B}{A numeric value, the number of permutations to be performed}

\item{alternative}{A character string specifying the alternative hypothesis.
Must be one of "two.sided" (default), "greater" or "less".}

\item{rowTestFUN}{A (vectorized) test function used in the two-sample case.
Defaults to \code{\link{rowWelchTests}}}

\item{rand.p.value}{A boolean value: should randomization \eqn{p}-values be
calculated and returned? Defaults to @FALSE}

\item{seed}{An integer (or NULL) value used as a seed for random number
generation. If \code{NULL}, no seed is specified}
}
\value{
a list with elements: \describe{

  \item{T}{A vector of \eqn{m} test statistics}

  \item{T0}{A \eqn{m \times B} matrix of randomized test statistics}

  \item{p}{A vector of \eqn{m} parametric \eqn{p}-values}

  \item{p0}{A \eqn{m \times B} matrix of parametric \eqn{p}-values on
  randomized data}

  \item{flavor}{A character value, the type of randomization performed:
  "perm" for permutation-based randomization in two-sample tests, and "flip"
  for sign-flipping-based randomization in one sample tests. See Details.}

  \item{rand.p}{A vector of \eqn{m} \eqn{p}-values (only if
  \code{rand.p.value} is \code{TRUE} )}

  \item{rand}{A \eqn{m \times B} matrix of randomization \eqn{p}-values (only
  if \code{rand.p.value} is \code{TRUE} )}
}
}
\description{
Randomization-based testing using permutation or sign-flipping
}
\details{
The type of randomization is determined by the column names of
  \code{X}. If these column names have exactly two distinct values, the
  corresponding columns are interpreted as two samples and a two-sample
  permutation-based test  is performed (flavor "perm"). Otherwise (including
  if \code{X} does not have column names), a one-sample test is performed
  using sign-flipping (flavor "flip").

  For permutation, we test the null hypothesis: "both groups have the same
  mean" against the alternative specified by parameter \code{alternative}. By
  default, the test is Welch's two-sample test for unequal variances, but
  other tests may be used via the argument \code{rowTestFUN}.  Permuted test
  statistics are calculated by B permutations of the group labels.
  Corresponding observed and permuted p-values are calculated as the
  proportion of permutations (including the identity) for which the permuted
  test statistic is larger than the observed test statistic.

  For sign-flipping, we test the null hypothesis: "the mean is 0" against the
  alternative specified by parameter \code{alternative}. We use the
  (rescaled) empirical mean of the observations as a test statistic.
  Sign-flipped test statistics are calculated by flipping the sign of each
  observation with probability 1/2.
}
\examples{

m <- 123
rho <- 0.2
n <- 100
pi0 <- 0.5
B <- 1e3

## two-sample data
sim <- gaussianSamples(m, rho, n, pi0, SNR = 2, prob = 0.5)
tests <- testByRandomization(sim$X, B)
testsW <- testByRandomization(sim$X, B = 10, rowTestFUN = rowWilcoxonTests)

## show test statistics
pch <- 20
colStat <- 1+sim$H
plot(tests$T, col = colStat, main = "T-Test statistics", pch = pch)
legend("topleft", c("H0", "H1"), pch=pch, col=1:2)

plot(testsW$T, col = colStat, main = "Wilcoxon test statistics", pch = pch)
legend("topleft", c("H0", "H1"), pch=pch, col=1:2)

# one-sample data:
sim <- gaussianSamples(m, rho, n, pi0, SNR=2)
tests <- testByRandomization(sim$X, B, alternative = "two.sided")

## show test statistics
pch <- 20
colStat <- 1+sim$H
plot(tests$T, col = colStat, main = "Test statistics", pch = pch)
legend("topleft", c("H0", "H1"), pch = pch, col = 1:2)

plot(-log10(tests$p), col = colStat, main = "-log[10](p-value)", pch = pch)
legend("topleft", c("H0", "H1"), pch = pch, col = 1:2)

}
\references{
Ge, Y., Dudoit, S. and Speed, T.P., 2003. Resampling-based
  multiple testing for microarray data analysis. _Test_, 12(1), pp.1-77.
}
