% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoGroupPermutationTest.R
\name{twoGroupPermutationTest}
\alias{twoGroupPermutationTest}
\title{Perform permutation testing for a two-group model}
\usage{
twoGroupPermutationTest(X, cls, B, seed = NULL)
}
\arguments{
\item{cls}{a vector of length \code{n} class labels in \code{0,1}}

\item{B}{a numeric values, the number of permutations to be performed}

\item{seed}{an integer (or NULL) value used as a seed for random number 
generation. If \code{NULL}, no seed is specified.}

\item{x}{a matrix of \code{m} variables by \code{n} observations}
}
\value{
a list with two
}
\description{
Perform permutation testing for a two-group model
}
\examples{

data(golub, package="multtest")
B <- 10000
  
}
\references{
Ge, Y., Dudoit, S. and Speed, T.P., 2003. Resampling-based
  multiple testing for microarray data analysis. _Test_, 12(1), pp.1-77.
}
