% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_functions.R
\name{V.star}
\alias{V.star}
\title{Post hoc bound on the number of false positives}
\usage{
V.star(S, C, ZL, leaf_list)
}
\arguments{
\item{S}{A subset of hypotheses}

\item{C}{Tree}

\item{ZL}{Zeta tree}

\item{leaf_list}{List of leaves}
}
\value{
An integer value, upper bound on the number false positives in S
}
\description{
Post hoc bound on the number of false positives
}
\references{
Durand, G., Blanchard, G., Neuvial, P., & Roquain, E. (2020). Post hoc false positive control for structured hypotheses. Scandinavian Journal of Statistics, 47(4), 1114-1148.
}
