% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot.numeric}
\alias{volcanoPlot.numeric}
\title{Volcano plot}
\usage{
\method{volcanoPlot}{numeric}(
  x,
  fc,
  thr,
  p = 1,
  q = 1,
  r = 0,
  cex = c(0.2, 0.6),
  col = c("#33333333", "#FF0000", "#FF666633"),
  pch = 19,
  ylim = NULL,
  ...
)
}
\arguments{
\item{x}{A vector of p-values}

\item{fc}{A vector of fold changes, of the same length as \code{x}}

\item{thr}{A numeric vector of length K, a JER controlling family}

\item{p}{A numeric value, the p-value threshold under which genes are selected}

\item{q}{A numeric value, the q-value (or FDR-adjusted p-value) threshold under which genes are selected}

\item{r}{A numeric value, the absolute fold change above which genes are selected}

\item{cex}{A numeric vector of length 2, the relative magnification factor for unselected (\code{cex[1]}) and unselected (\code{cex[2]}) genes.}

\item{col}{A vector of length 3}

\item{pch}{An integer or single character string specifying the plotting character, see \code{\link{par}}}

\item{ylim}{A numeric vector of length 2, the \eqn{y} limits of the plot}

\item{...}{Not used}
}
\value{
The indices of selected genes (returned invisibly)
}
\description{
Volcano plot for differential expression studies
}
\details{
A Welch T-test of differential expression between the two categories
defined by \code{categ} are applied for each gene using the
\code{\link{rowWelchTests}} function, which also outputs the "fold change"
(mean difference in log scale) between the two categories.
}
\seealso{
Volcano plot shiny app at \url{https://pneuvial.shinyapps.io/volcano-plot/}
}
