% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Volcano plot}
\usage{
volcanoPlot(x, ...)
}
\arguments{
\item{x}{An object. See individual methods for specifics}

\item{...}{Other arguments passed to methods}
}
\description{
Volcano plot
}
