% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Volcano plot}
\usage{
volcanoPlot(
  X,
  categ,
  thr,
  p = 1,
  q = 1,
  r = 0,
  cex = c(0.2, 0.6),
  col = c("#33333333", "#FF0000", "#FF666633"),
  pch = 19,
  ylim = NULL
)
}
\arguments{
\item{X}{A matrix of \eqn{m} variables (hypotheses) by \eqn{n} observations}

\item{categ}{A numeric vector of \code{n} categories in \eqn{0, 1} for the
observations}

\item{thr}{A numeric vector of length K, a JER controlling family}

\item{p}{A numeric value, the p-value threshold under which genes are selected}

\item{q}{A numeric value, the q-value (or FDR-adjusted p-value) threshold under which genes are selected}

\item{r}{A numeric value, the absolute fold change above which genes are selected}

\item{cex}{A numeric vector of length 2, the relative magnification factor for unselected (\code{cex[1]}) and unselected (\code{cex[2]}) genes.}

\item{col}{A vector of length 3}

\item{pch}{An integer or single character string specifying the plotting character, see \code{\link{par}}}

\item{ylim}{A numeric vector of length 2, the \eqn{y} limits of the plot}
}
\value{
The indices of selected genes (returned invisibly)
}
\description{
Volcano plot for differential expression studies
}
\details{
A Welch T-test of differential expression between the two categories
  defined by \code{categ} are applied for each gene using the
  \code{\link{rowWelchTests}} function, which also outputs the "fold change"
  (mean difference in log scale) between the two categories.
}
\examples{
m <- 500
pi0 <- 0.5
m1 <- m-m*pi0
SNR <- 5*(runif(m1)-0.5)
sim <- gaussianSamples(m = m, rho = 0.2, n = 100,
                       pi0 = pi0, SNR = SNR, prob = 0.5)
X <- sim$X
categ <- sim$categ
alpha <- 0.2
cal <- calibrateJER(X = X, categ = categ, B = 1e2, alpha = alpha, refFamily="Simes")
sel <- volcanoPlot(X = X, categ = categ, thr = cal$thr, q = 0.2, r = 0.2, ylim = c(0, 6))

# Compare bound to reality
TP <- sum(sim$H[sel])
FP <- sum(1-sim$H[sel])
FDP <- FP/length(sel)
}
\seealso{
Volcano plot shiny app at \url{https://pneuvial.shinyapps.io/volcano-plot/}
}
