% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SansSouci-class.R
\name{volcanoPlot.SansSouci}
\alias{volcanoPlot.SansSouci}
\title{Volcano plot for a \code{SansSouci} object}
\usage{
\method{volcanoPlot}{SansSouci}(
  x,
  fold_change = foldChanges(x)[contrast_name, ],
  p_value = pValues(x)[contrast_name, ],
  p = 1,
  q = 1,
  r = 0,
  contrast_name = x$input$contrast_name[1],
  cex = c(0.4, 1.5),
  col = c("#33333333", "#FF0000", "#FF666633"),
  pch = 19,
  feature_label = "feature",
  ylim = NULL,
  interactive = FALSE,
  ...
)
}
\arguments{
\item{x}{A numeric vector of fold changes (x axis of the volcano plot)}

\item{fold_change}{An optional vector of fold changes, of the same length as \code{nHyp(x)}, used for volcanoPlot x-axis. If not specified, \code{foldChanges(x)} is used.}

\item{p_value}{A vector of p-values, of the same length as \code{nHyp(x)}, used for volcanoPlot y-axis. If not specified, \code{pValues(x)} is used}

\item{p}{A numeric value, the p-value threshold under which features are selected}

\item{q}{A numeric value, the q-value (or FDR-adjusted p-value) threshold under which features are selected}

\item{r}{A numeric value, the absolute fold change above which features are selected}

\item{contrast_name}{A character value, the selected contrast. Should be chosen in \code{x$input$contrast_name}.}

\item{cex}{A numeric vector of length 2, the relative magnification factor for unselected (\code{cex[1]}) and unselected (\code{cex[2]}) features.}

\item{col}{A vector of length 3}

\item{pch}{An integer or single character string specifying the plotting character, see \code{\link{par}}}

\item{feature_label}{A character, the label to be used to designate individual hypotheses in the plot title
(e.g. "gene", "proteins", ...)}

\item{ylim}{A numeric vector of length 2, the \eqn{y} limits of the plot}

\item{interactive}{logical. If TRUE, a plotly object is returned then a
ggplot2 object is returned.}

\item{...}{Not used}
}
\description{
Volcano plot for a \code{SansSouci} object
}
\details{
The default is to use the fold changes and p-values from the input SansSouci object. See the vignette
\url{https://sanssouci-org.github.io/sanssouci/articles/post-hoc_differential-expression_RNAseq.html#custom-statistics-example-using-limma-voom}
for an example where custom fold changes and p-values are used.
}
\examples{
data(expr_ALL, package = "sanssouci.data")
groups <- ifelse(colnames(expr_ALL) == "NEG", 0, 1)
a <- SansSouci(Y = expr_ALL, groups = groups)

res <- fit(a, B = 100, alpha = 0.1)
volcanoPlot(res, q = 0.2, r = 0.2, ylim = c(0, 4), feature_label = "gene")
}
