% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-functions.R
\name{zetas.tree}
\alias{zetas.tree}
\alias{zetas.tree.refined}
\title{Estimate of the proportion of true nulls in each node of a tree}
\usage{
zetas.tree(C, leaf_list, method, pvalues, alpha)

zetas.tree.refined(C, leaf_list, method, pvalues, alpha)
}
\arguments{
\item{C}{Tree structure}

\item{leaf_list}{A list of leaves}

\item{method}{A function to compute the estimators}

\item{pvalues}{A vector of \eqn{p}-values}

\item{alpha}{A target level}
}
\description{
Estimate of the proportion of true nulls in each node of a tree
}
\details{
The proportion of true nulls in each node is estimated by an union bound on the regions. That is, the provided method is applied at level \code{alpha/nR} where \code{nR} is the number of regions.

In \code{zetas.tree.refined}, one tries to estimate the number of regions containing only signal.
}
\examples{

m <- 1000 
dd <- dyadic.from.window.size(m, s = 10, method = 2)
leaf_list <- dd$leaf_list
mu <- gen.mu.leaves(m, K1 = 3, d = 1, grouped = FALSE, "const", barmu = 4, leaf_list)
pvalues<-gen.p.values(m, mu, rho = 0)
C <- dd$C 
ZL<-zetas.tree(C, leaf_list, zeta.DKWM, pvalues, alpha = 0.05)
}
\references{
Durand, G., Blanchard, G., Neuvial, P., & Roquain, E. (2020). Post hoc false positive control for structured hypotheses. Scandinavian Journal of Statistics, 47(4), 1114-1148.
}
